% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{A01}
\alias{A01}
\title{Check one or more a well names to make sure they follow the A01 or A:1 format}
\usage{
A01(well, form = c("A:1", "A01")[1])
}
\arguments{
\item{well}{character vector of well names such as A:1 or A01}

\item{form}{character, the desired output format either "A:1" or "A01"}
}
\value{
a character vector of well names formatted as specified
}
\description{
Check one or more a well names to make sure they follow the A01 or A:1 format
}

