% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artifact.R
\name{artifact_get_root}
\alias{artifact_get_root}
\title{Given an ArtifactRefClass find it's root artifact (submitted sample) by 
searching parent-processes until there are none to search.}
\usage{
artifact_get_root(x)
}
\arguments{
\item{x}{ArtifactRefClass object}
}
\value{
ArtifactRefClass or NULL
}
\description{
Given an ArtifactRefClass find it's root artifact (submitted sample) by 
searching parent-processes until there are none to search.
}
