% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ArtifactGroup.R
\name{ArtifactGroupRefClass_get_artifacts}
\alias{ArtifactGroupRefClass_get_artifacts}
\title{Retrieve a list ArtifactRefClass objects}
\value{
a named vector of ArtifactRefClass objects
}
\description{
Retrieve a list ArtifactRefClass objects
}
\seealso{
Other ArtifactGroup: \code{\link{ArtifactGroupRefNode_POST}};
  \code{\link{ArtifactGroupRefNode_PUT}};
  \code{\link{ArtifactGrouptRefNode_DELETE}}
}

