% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ArtifactGroup.R
\name{ArtifactGroupRefNode_DELETE}
\alias{ArtifactGroupRefNode_DELETE}
\title{DELETE is disallowed for ArtifactGroup}
\description{
DELETE is disallowed for ArtifactGroup
}
\seealso{
Other ArtifactGroup: \code{\link{ArtifactGroupRefClass_get_artifacts}};
  \code{\link{ArtifactGroupRefNode_POST}};
  \code{\link{ArtifactGroupRefNode_PUT}}
}

