% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArtifactGroup.R
\name{ArtifactGroupRefNode_POST}
\alias{ArtifactGroupRefNode_POST}
\title{POST is disallowed for artifact groups}
\description{
POST is disallowed for artifact groups
}
\seealso{
Other ArtifactGroup: \code{\link{ArtifactGroupRefClass_get_artifacts}},
  \code{\link{ArtifactGroupRefNode_DELETE}},
  \code{\link{ArtifactGroupRefNode_PUT}}
}
