% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ArtifactGroup.R
\name{ArtifactGroupRefNode_PUT}
\alias{ArtifactGroupRefNode_PUT}
\title{PUT is disallowed for artifact groups}
\description{
PUT is disallowed for artifact groups
}
\seealso{
Other ArtifactGroup: \code{\link{ArtifactGroupRefClass_get_artifacts}};
  \code{\link{ArtifactGroupRefNode_DELETE}};
  \code{\link{ArtifactGroupRefNode_POST}}
}

