% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Artifact.R
\name{ArtifactRefClass_get_artifact_group}
\alias{ArtifactRefClass_get_artifact_group}
\title{Get the artifact group assocated with this artifact}
\arguments{
\item{form}{character flag of type to return "Node" or "uri" or "name"}
}
\value{
XML::xmlNode (or NULL) or the uri (or "")
}
\description{
Get the artifact group assocated with this artifact
}

