% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artifact.R
\name{ArtifactRefClass_get_artifact_groups}
\alias{ArtifactRefClass_get_artifact_groups}
\title{Get the artifact groups associated with this artifact if any}
\arguments{
\item{form}{character flag of type to return "Node" or "uri" or "name"}
}
\value{
list of ArtifactGroupRefClass (or NULL) or vector of names or uris (or "")
}
\description{
Get the artifact groups associated with this artifact if any
}
