% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Artifact.R
\name{ArtifactRefClass_get_container}
\alias{ArtifactRefClass_get_container}
\title{Get the artifact container as uri or NodeRefClass}
\arguments{
\item{form}{character either "uri" or "Node"}
}
\value{
ContainerRefClass or uri (or NULL or "")
}
\description{
Get the artifact container as uri or NodeRefClass
}
\seealso{
Other Artifact: \code{\link{ArtifactRefClass_get_parent_process}};
  \code{\link{ArtifactRefClass_get_sample}};
  \code{\link{ArtifactRefClass_get_wellname}};
  \code{\link{ArtifactRefClass_location}};
  \code{\link{ArtifactRefClass_qc_flag}};
  \code{\link{ArtifactRefClass_working_flag}};
  \code{\link{ArtifactRefNode_DELETE}};
  \code{\link{ArtifactRefNode_POST}}
}

