% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artifact.R
\name{ArtifactRefClass_get_file_artifact}
\alias{ArtifactRefClass_get_file_artifact}
\title{Retrieve the artifact's file artifact (if any)}
\arguments{
\item{form}{character flag of type to return "Node" or "uri"}
}
\value{
FileRefClass (or NULL) or the uri (or "")
}
\description{
Retrieve the artifact's file artifact (if any)
}
\seealso{
Other Artifact: \code{\link{ArtifactRefClass_get_container}},
  \code{\link{ArtifactRefClass_get_parent_process}},
  \code{\link{ArtifactRefClass_get_samples}},
  \code{\link{ArtifactRefClass_get_sample}},
  \code{\link{ArtifactRefClass_get_wellname}},
  \code{\link{ArtifactRefClass_location}},
  \code{\link{ArtifactRefClass_qc_flag}},
  \code{\link{ArtifactRefClass_working_flag}},
  \code{\link{ArtifactRefNode_DELETE}},
  \code{\link{ArtifactRefNode_POST}}
}
