% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artifact.R
\name{ArtifactRefClass_get_parent_process}
\alias{ArtifactRefClass_get_parent_process}
\title{Get the parent_process assocated with this artifact}
\arguments{
\item{form}{character flag of type to return "Node" or "uri"}
}
\value{
XML::xmlNode (or NULL) or the uri (or "")
}
\description{
Get the parent_process assocated with this artifact
}
\seealso{
Other Artifact: \code{\link{ArtifactRefClass_get_container}},
  \code{\link{ArtifactRefClass_get_file_artifact}},
  \code{\link{ArtifactRefClass_get_sample}},
  \code{\link{ArtifactRefClass_get_wellname}},
  \code{\link{ArtifactRefClass_location}},
  \code{\link{ArtifactRefClass_qc_flag}},
  \code{\link{ArtifactRefClass_working_flag}},
  \code{\link{ArtifactRefNode_DELETE}},
  \code{\link{ArtifactRefNode_POST}}
}
