% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artifact.R
\name{ArtifactRefNode_DELETE}
\alias{ArtifactRefNode_DELETE}
\title{DELETE is disallowed for artifacts}
\description{
DELETE is disallowed for artifacts
}
\seealso{
Other Artifact: \code{\link{ArtifactRefClass_get_container}},
  \code{\link{ArtifactRefClass_get_file_artifact}},
  \code{\link{ArtifactRefClass_get_parent_process}},
  \code{\link{ArtifactRefClass_get_sample}},
  \code{\link{ArtifactRefClass_get_wellname}},
  \code{\link{ArtifactRefClass_location}},
  \code{\link{ArtifactRefClass_qc_flag}},
  \code{\link{ArtifactRefClass_working_flag}},
  \code{\link{ArtifactRefNode_POST}}
}
