% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Udf.R
\name{as_ascii}
\alias{as_ascii}
\title{Sanitize a character vector to be UTF-8 ASCII}
\usage{
as_ascii(x, sub = "")
}
\arguments{
\item{x}{character vector}

\item{sub}{the character used to replace non-ASCII}
}
\value{
character vector
}
\description{
Sanitize a character vector to be UTF-8 ASCII
}

