% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{batch_create}
\alias{batch_create}
\title{Create one or more nodes (Sample and Container only)}
\usage{
batch_create(x, lims, asNode = asNode, rel = c("samples", "containers")[1])
}
\arguments{
\item{x}{a list of one or more XML::xmlNode}

\item{lims}{a LimsRefClass node}

\item{asNode}{logical, if TRUE return a named list of NodeRefClass objects}

\item{rel}{the relative namespace into the "batch/create"}
}
\value{
list of XML::xmlNode or NodeRefClass
}
\description{
Return order is *not* enforced to be the same as the input order
}

