% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{build_config_path}
\alias{build_config_path}
\title{Create a configuration file name in the form of \code{/path/to/.hostname-rc}}
\usage{
build_config_path(hostname = system("hostname", intern = TRUE), path = "~")
}
\arguments{
\item{hostname}{character hostname}

\item{path}{character the directory of the configuration file}
}
\value{
the complete configuration file path
}
\description{
Create a configuration file name in the form of \code{/path/to/.hostname-rc}
}
\examples{
 \dontrun{
 build_config_path()
 [1] "~/.scgc-clarity-dev-rc"
 build_config_path(hostname = 'foo')
 [1] "~/.foo-rc"
 build_config_path(hostname = 'foo', path = '/stairway/to/heaven')
 [1] "/stairway/to/heaven/.foo-rc"
 build_config_path(hostname = 'foo.bar.org', path = '/stairway/to/heaven')
 [1] "/stairway/to/heaven/.foo-rc"
 }
}

