% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{build_query}
\alias{build_query}
\title{Prepares a contribution to a query list - see \code{modify_url}}
\usage{
build_query(x)
}
\arguments{
\item{x}{a named list like \code{list(name=c("boo", "foo"), type = "shoe")}}
}
\value{
character like "name=boo&name=foo$type=shoe"
}
\description{
Prepares a contribution to a query list - see \code{modify_url}
}

