% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Udf.R
\name{check_type}
\alias{check_type}
\title{Given a LIMS type code and value, cast the value to be LIMS friendly}
\usage{
check_type(typ, value)
}
\arguments{
\item{typ}{character LIMS type (Numeric, String, Text, etc)}

\item{value}{any value}
}
\value{
the value cast to the specified type
}
\description{
Given a LIMS type code and value, cast the value to be LIMS friendly
}
