% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefClass_get_artifacts}
\alias{ContainerRefClass_get_artifacts}
\title{Retrieve a named list of artifact XML:xmlNode or ArtifactRefClass objects}
\arguments{
\item{asNode}{logical if TURE cast the result to ArtifactRefClass}
}
\value{
a named vector of artifact XML:xmlNode or ArtifactRefClass objects
}
\description{
Retrieve a named list of artifact XML:xmlNode or ArtifactRefClass objects
}
\seealso{
Other Container: \code{\link{ContainerRefClass_POST}};
  \code{\link{ContainerRefClass_get_placements}};
  \code{\link{ContainerRefClass_get_samples}};
  \code{\link{ContainerRefClass_n_empty}};
  \code{\link{ContainerRefClass_n_occupied}};
  \code{\link{ContainerRefClass_n_wells}}
}

