% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefClass_get_placements}
\alias{ContainerRefClass_get_placements}
\title{Retrieve a named vector of placements uri}
\arguments{
\item{placement}{character vector of well names (A:1 etc.) or NULL to get all}
}
\value{
a named vector of placements, possibly an empty character vector
}
\description{
Retrieve a named vector of placements uri
}
\seealso{
Other Container: \code{\link{ContainerRefClass_POST}},
  \code{\link{ContainerRefClass_get_artifacts}},
  \code{\link{ContainerRefClass_get_samples}},
  \code{\link{ContainerRefClass_get_type}},
  \code{\link{ContainerRefClass_n_empty}},
  \code{\link{ContainerRefClass_n_occupied}},
  \code{\link{ContainerRefClass_n_wells}},
  \code{\link{ContainerRefClass_required_namespaces}},
  \code{\link{ContainerRefClass_well_occupied}}
}
