% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefClass_get_type}
\alias{ContainerRefClass_get_type}
\title{Retrieve a named vector of container type attributes (uri and name)}
\value{
a named vector of uri and name, or NULL
}
\description{
Retrieve a named vector of container type attributes (uri and name)
}
\seealso{
Other Container: \code{\link{ContainerRefClass_POST}},
  \code{\link{ContainerRefClass_get_artifacts}},
  \code{\link{ContainerRefClass_get_placements}},
  \code{\link{ContainerRefClass_get_samples}},
  \code{\link{ContainerRefClass_n_empty}},
  \code{\link{ContainerRefClass_n_occupied}},
  \code{\link{ContainerRefClass_n_wells}},
  \code{\link{ContainerRefClass_required_namespaces}},
  \code{\link{ContainerRefClass_well_occupied}}
}

