% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefClass_n_occupied}
\alias{ContainerRefClass_n_occupied}
\title{Retrieve the occupied well count}
\value{
numeric
}
\description{
Retrieve the occupied well count
}
\seealso{
Other Container: \code{\link{ContainerRefClass_POST}};
  \code{\link{ContainerRefClass_get_artifacts}};
  \code{\link{ContainerRefClass_get_placements}};
  \code{\link{ContainerRefClass_get_samples}};
  \code{\link{ContainerRefClass_n_empty}};
  \code{\link{ContainerRefClass_n_wells}};
  \code{\link{ContainerRefClass_required_namespaces}}
}

