% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefClass_POST}
\alias{ContainerRefClass_POST}
\title{POST is disallowed for containers}
\description{
POST is disallowed for containers
}
\seealso{
Other Container: \code{\link{ContainerRefClass_get_artifacts}};
  \code{\link{ContainerRefClass_get_placements}};
  \code{\link{ContainerRefClass_get_samples}};
  \code{\link{ContainerRefClass_n_empty}};
  \code{\link{ContainerRefClass_n_occupied}};
  \code{\link{ContainerRefClass_n_wells}}
}

