% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefClass_required_namespaces}
\alias{ContainerRefClass_required_namespaces}
\title{Return a character vector of required namespace names}
\value{
character vector
}
\description{
Return a character vector of required namespace names
}
\seealso{
Other Container: \code{\link{ContainerRefClass_POST}},
  \code{\link{ContainerRefClass_get_artifacts}},
  \code{\link{ContainerRefClass_get_placements}},
  \code{\link{ContainerRefClass_get_samples}},
  \code{\link{ContainerRefClass_get_type}},
  \code{\link{ContainerRefClass_n_empty}},
  \code{\link{ContainerRefClass_n_occupied}},
  \code{\link{ContainerRefClass_n_wells}},
  \code{\link{ContainerRefClass_well_occupied}}
}

