% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefClass_well_occupied}
\alias{ContainerRefClass_well_occupied}
\title{Test if one or more wells are occupied}
\arguments{
\item{placement}{character vector of well names}
}
\value{
logical named vector
}
\description{
Test if one or more wells are occupied
}
\seealso{
Other Container: \code{\link{ContainerRefClass_POST}},
  \code{\link{ContainerRefClass_get_artifacts}},
  \code{\link{ContainerRefClass_get_placements}},
  \code{\link{ContainerRefClass_get_samples}},
  \code{\link{ContainerRefClass_get_type}},
  \code{\link{ContainerRefClass_n_empty}},
  \code{\link{ContainerRefClass_n_occupied}},
  \code{\link{ContainerRefClass_n_wells}},
  \code{\link{ContainerRefClass_required_namespaces}}
}

