% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefNode_get_placements}
\alias{ContainerRefNode_get_placements}
\title{Retrieve a named vector of placements uri}
\value{
a named vector of placements, possibly an empty charcater vector
}
\description{
Retrieve a named vector of placements uri
}
\seealso{
Other Container: \code{\link{ContainerRefNode_POST}};
  \code{\link{ContainerRefNode_get_artifacts}};
  \code{\link{ContainerRefNode_get_samples}};
  \code{\link{ContainerRefNode_n_empty}};
  \code{\link{ContainerRefNode_n_occupied}};
  \code{\link{ContainerRefNode_n_wells}}
}

