% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefNode_get_samples}
\alias{ContainerRefNode_get_samples}
\title{Retrieve a named list of artifact XML:xmlNode or SampleRefClass}
\arguments{
\item{asNode}{logical if TURE cast the result to SampleRefClass}
}
\value{
a named vector of sample XML:xmlNode or SampleRefClass objects
}
\description{
Retrieve a named list of artifact XML:xmlNode or SampleRefClass
}
\seealso{
Other Container: \code{\link{ContainerRefNode_POST}};
  \code{\link{ContainerRefNode_get_artifacts}};
  \code{\link{ContainerRefNode_get_placements}};
  \code{\link{ContainerRefNode_n_empty}};
  \code{\link{ContainerRefNode_n_occupied}};
  \code{\link{ContainerRefNode_n_wells}}
}

