% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefNode_n_occupied}
\alias{ContainerRefNode_n_occupied}
\title{Retrieve the occupied well count}
\value{
numeric
}
\description{
Retrieve the occupied well count
}
\seealso{
Other Container: \code{\link{ContainerRefNode_POST}};
  \code{\link{ContainerRefNode_get_artifacts}};
  \code{\link{ContainerRefNode_get_placements}};
  \code{\link{ContainerRefNode_get_samples}};
  \code{\link{ContainerRefNode_n_empty}};
  \code{\link{ContainerRefNode_n_wells}}
}

