% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefNode_POST}
\alias{ContainerRefNode_POST}
\title{POST is disallowed for containers}
\description{
POST is disallowed for containers
}
\seealso{
Other Container: \code{\link{ContainerRefNode_get_artifacts}};
  \code{\link{ContainerRefNode_get_placements}};
  \code{\link{ContainerRefNode_get_samples}};
  \code{\link{ContainerRefNode_n_empty}};
  \code{\link{ContainerRefNode_n_occupied}};
  \code{\link{ContainerRefNode_n_wells}}
}

