% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Containertype.R
\docType{class}
\name{ContainerTypeRefClass-class}
\alias{ContainerTypeRefClass}
\alias{ContainerTypeRefClass-class}
\title{A Containertype representation that subclasses from NodeRefClass}
\description{
A Containertype representation that subclasses from NodeRefClass
}
\section{Fields}{

\describe{
\item{\code{name}}{character, name if any}

\item{\code{is_tube}}{logical, type if any}

\item{\code{xdim,}}{list [is_alpha = TRUE/FALSE, offset = 0/1, size = n]}

\item{\code{ydim,}}{list [is_alpha = TRUE/FALSE, offset = 0/1, size = n]}
}}
\seealso{
Other Node: \code{\link{ArtifactGroupRefClass}},
  \code{\link{ArtifactGroupRefClass-class}};
  \code{\link{ArtifactRefClass}},
  \code{\link{ArtifactRefClass-class}};
  \code{\link{ContainerRefClass}},
  \code{\link{ContainerRefClass-class}};
  \code{\link{FieldRefClass_show}};
  \code{\link{FieldRefClass}},
  \code{\link{FieldRefClass-class}};
  \code{\link{FileRefClass}},
  \code{\link{FileRefClass-class}};
  \code{\link{InputOutputMapRefClass}},
  \code{\link{InputOutputMapRefClass-class}};
  \code{\link{NodeRefClass_BROWSE}};
  \code{\link{NodeRefClass_DELETE}};
  \code{\link{NodeRefClass_GET}};
  \code{\link{NodeRefClass_POST}};
  \code{\link{NodeRefClass_PUT}};
  \code{\link{NodeRefClass_get_childv}};
  \code{\link{NodeRefClass_get_field}};
  \code{\link{NodeRefClass_get_type}},
  \code{\link{NodeRefClass_get_type}};
  \code{\link{NodeRefClass_has_child}};
  \code{\link{NodeRefClass_has_lims}};
  \code{\link{NodeRefClass_new}};
  \code{\link{NodeRefClass_set_udfs}};
  \code{\link{NodeRefClass_show}};
  \code{\link{NodeRefClass_toString}};
  \code{\link{NodeRefClass_unames}};
  \code{\link{NodeRefClass_update}};
  \code{\link{NodeRefClass}},
  \code{\link{NodeRefClass-class}};
  \code{\link{ProcessRefClass}},
  \code{\link{ProcessRefClass-class}};
  \code{\link{ProjectRefClass}},
  \code{\link{ProjectRefClass-class}};
  \code{\link{ProjectsRefClass}},
  \code{\link{ProjectsRefClass-class}};
  \code{\link{ResearcherRefClass}},
  \code{\link{ResearcherRefClass-class}};
  \code{\link{SampleRefClass}},
  \code{\link{SampleRefClass-class}};
  \code{\link{udf_type2R}}
}

