% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file_transfer.R
\name{cp_download}
\alias{cp_download}
\title{Download files using system cp}
\usage{
cp_download(url, dest, verbose = FALSE, extra = "")
}
\arguments{
\item{url}{character}

\item{dest}{character destination filename, by default the basename of the URL}

\item{verbose}{logical, if TRUE then echo the command string}

\item{extra}{character extra params for cp, currently ""}
}
\value{
integer with 0 for success
}
\description{
This assumes permissions have been set up
}
\seealso{
Other file_transfer: \code{\link{duck_download}};
  \code{\link{duck_upload}}; \code{\link{has_duck}};
  \code{\link{has_rsync}}; \code{\link{has_scp}};
  \code{\link{rsync_download}}; \code{\link{rsync_upload}};
  \code{\link{scp_download}}; \code{\link{scp_upload}}
}

