% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artifact.R
\name{create_artifacts_details}
\alias{create_artifacts_details}
\title{Create a artifacts details node assembled from one or more artifacts XML::xmlNode
See \url{http://genologics.com/files/permanent/API/latest/data_art.html#element_details}}
\usage{
create_artifacts_details(x)
}
\arguments{
\item{x}{one or more XML::xmlNode for artifacts or ArtifactRefClass objects}

\item{a}{artifacts details XML:xmlNode node}
}
\description{
Create a artifacts details node assembled from one or more artifacts XML::xmlNode
See \url{http://genologics.com/files/permanent/API/latest/data_art.html#element_details}
}

