% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Artifact.R
\name{create_artifacts_links}
\alias{create_artifacts_links}
\title{Create an artifacts links node from uris suitable for batch operations
See \url{http://genologics.com/files/permanent/API/latest/rest.version.artifacts.batch.retrieve.html}}
\usage{
create_artifacts_links(x)
}
\arguments{
\item{x}{one or more URI}
}
\value{
a links node
}
\description{
Create an artifacts links node from uris suitable for batch operations
See \url{http://genologics.com/files/permanent/API/latest/rest.version.artifacts.batch.retrieve.html}
}

