% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container.R
\name{create_container_node}
\alias{create_container_node}
\title{Create a container XML::xmlNode suitable for POSTing}
\usage{
create_container_node(type_uri, name = NULL, state = "Empty")
}
\arguments{
\item{type_uri}{character uri of container type (required)}

\item{name}{character container name (optional)}

\item{state}{character generally Empty but possibly Populated, Discarded, or Depleted}
}
\value{
XML::xmlNode
}
\description{
Create a container XML::xmlNode suitable for POSTing
}
\seealso{
Other Lims Container: \code{\link{LimsRefClass_get_containers}},
  \code{\link{LimsRefClass_get_containertypes}}
}

