% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Container.R
\name{create_containers_details}
\alias{create_containers_details}
\title{Create a containers details node assembled from one or more container XML::xmlNode
See \url{http://genologics.com/files/permanent/API/latest/data_con.html#element_details}}
\usage{
create_containers_details(x, form = c("retrieve", "update", "create"))
}
\arguments{
\item{x}{one or more XML::xmlNode for container or ContainerRefClass objects}

\item{form}{character}

\item{a}{containers details XML:xmlNode node}
}
\description{
Create a containers details node assembled from one or more container XML::xmlNode
See \url{http://genologics.com/files/permanent/API/latest/data_con.html#element_details}
}

