% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{create_file_node}
\alias{create_file_node}
\title{Create an unresolved file node}
\usage{
create_file_node(attached_to = "", original_location = "",
  namespace = "file")
}
\arguments{
\item{attached_to}{character uri of the artifact to attach the file to}

\item{original_location}{character, the fully qualified path of the original file}

\item{namespace}{character the namespace for the resource}
}
\value{
XML::xmlNode
}
\description{
Create an unresolved file node
}

