% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputOutputMap.R
\name{create_iom_analyte}
\alias{create_iom_analyte}
\title{Create an 'input-output-map' node for Analyte - useful when 
programmatically running a process.}
\usage{
create_iom_analyte(inputartifacturi, container_uri = "container_uri",
  placement = "A:1", output_type = "Analyte")
}
\arguments{
\item{inputartifacturi}{character vector of input artifact URIs}

\item{container_uri}{character vector, the uri of the destination container}

\item{placement}{character vector, the 'A:1' form well placement}

\item{output_type}{character, you shouldn't have to change this}
}
\value{
an XML::xmlNode of type input-output-map for 'Analyte'
}
\description{
Create an 'input-output-map' node for Analyte - useful when 
programmatically running a process.
}
\seealso{
\url{https://genologics.zendesk.com/entries/23659973-Running-a-Process}
}
