% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputOutputMap.R
\name{create_iom_shared_resultfile}
\alias{create_iom_shared_resultfile}
\title{Create an 'input-output-map' node for a shared result file - useful when 
programmatically running a process. Multiple inputs = one output}
\usage{
create_iom_shared_resultfile(inputartifacturi, output_type = "ResultFile")
}
\arguments{
\item{inputartifacturi}{character vector of URIs}

\item{output_type}{character, you shouldn't have to change this}
}
\value{
an XML::xmlNode of type input-output-artifact
}
\description{
Create an 'input-output-map' node for a shared result file - useful when 
programmatically running a process. Multiple inputs = one output
}
\seealso{
\url{https://genologics.zendesk.com/entries/23659973-Running-a-Process}
}
