% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab.R
\name{create_lab_node}
\alias{create_lab_node}
\title{Create a lab/account XML::xmlNode suitable for POSTing}
\usage{
create_lab_node(name, shipping_address = NULL, billing_address = NULL,
  website = NULL)
}
\arguments{
\item{name}{character lab/account name (required)}

\item{shipping_address}{named character vector ala
[street='123 Main', city='East Boothbay Harbor', etc.]}

\item{billing_address}{named character vector}

\item{website}{character website}
}
\value{
XML::xmlNode
}
\description{
Create a lab/account XML::xmlNode suitable for POSTing
}

