% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{create_process_node}
\alias{create_process_node}
\title{Create a process node used for process creation}
\usage{
create_process_node(type = "processTypeName", technician = "researcherURI",
  dateRun = format(Sys.Date(), "\%Y-\%m-\%d"), instrument = NULL, ...)
}
\arguments{
\item{type}{character, the name of the process ("FACS-2", ...)}

\item{technician}{character,the uri of the technician}

\item{dateRun}{character, the date of the run}

\item{instrument}{character, optional, the URI of the instrument}

\item{...}{further arguments for \code{\link{newXMLNode}} including child nodes}
}
\value{
xmlNode of type 'process'
}
\description{
Create a process node used for process creation
}
\seealso{
Other Process: \code{\link{ProcessRefClass_date_run}},
  \code{\link{ProcessRefClass_get_artifacts}},
  \code{\link{ProcessRefClass_get_file_artifacts}},
  \code{\link{ProcessRefClass_get_parent_process}},
  \code{\link{ProcessRefClass_get_type}},
  \code{\link{ProcessRefClass_instrument}},
  \code{\link{ProcessRefClass_technician}}
}

