% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Researcher.R
\name{create_researcher_node}
\alias{create_researcher_node}
\title{Create a researcher XML::xmlNode suitable for POSTing}
\usage{
create_researcher_node(firstname = NULL, lastname = NULL, email = NULL,
  username = NULL, password = NULL, initials = NULL, role = NULL,
  account_locked = NULL, lab = NULL)
}
\arguments{
\item{firstname}{character researcher first name (required)}

\item{lastname}{character researcher last name (required)}

\item{email}{character researcher email (required)}

\item{username}{character if credentailed the required}

\item{password}{character if credentailed then required}

\item{initials}{exactly 3 alphanumeric characters, if credentialed then required}

\item{role}{character vector of one or more roles}

\item{account_locked}{logical}

\item{lab}{character URI of the lab}
}
\value{
XML::xmlNode
}
\description{
Create a researcher XML::xmlNode suitable for POSTing
}
\seealso{
Other Lims Researcher: \code{\link{LimsRefClass_get_researchers}}
}

