% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{create_routing}
\alias{create_routing}
\title{Create routing node}
\usage{
create_routing()
}
\value{
a routing node
}
\description{
Create routing node
}
\seealso{
\url{https://www.genologics.com/files/permanent/API/latest/data_rt.html#routing}
}
