% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{create_routing_artifacts}
\alias{create_routing_artifacts}
\title{Create or update an artifact routing map}
\usage{
create_routing_artifacts(x, workflow_uri = NULL, stage_uri = NULL,
  action = c("assign", "unassign")[1], routing = create_routing())
}
\arguments{
\item{x}{a list of one or more artifacts as ArtifactsRefClass or xmlNode}

\item{workflow_uri}{from the link... 'The workflow to use for (un)assigning 
the samples. Is not needed if stage URI is set. If stageURI is not 
specified, the default stage in the workflow will be used for assignment.'}

\item{action}{character of either 'assign' or 'unassign'}

\item{routing}{xml structure of type 'routing' - if NULL a new one is created
but if provided it is updated}

\item{satge_uri}{from the link above 'The stage to assign the samples to. The #'  stage is tied to a workflow, so workflowURI is not needed if this attribute #'  is set.'}
}
\value{
xml routing node (new or updated) or NULL
}
\description{
Create or update an artifact routing map
}
\seealso{
\url{https://www.genologics.com/files/permanent/API/latest/data_rt.html#element_routing}
}
