% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Sample.R
\name{create_sample_node}
\alias{create_sample_node}
\title{Create a sample XML::xmlNode suitable for POSTing}
\usage{
create_sample_node(name = NULL, project_uri = NULL, container_uri = NULL,
  well = NULL)
}
\arguments{
\item{name}{character container name (required)}

\item{project_uri}{character uri of project (required)}

\item{container_uri}{character uri of the container (required)}

\item{well}{character location on the well such as 'A:1' (required)}
}
\value{
XML::xmlNode
}
\description{
Create a sample XML::xmlNode suitable for POSTing
}
\seealso{
Other Lims Sample: \code{\link{LimsRefClass_get_samples}}
}

