% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.R
\name{create_samples_details}
\alias{create_samples_details}
\title{Create a sample details node assembled from one or more sample XML::xmlNode
See \url{http://genologics.com/files/permanent/API/latest/data_smp.html#element_details}}
\usage{
create_samples_details(x)
}
\arguments{
\item{x}{one or more XML::xmlNode for sample or SampleRefClass objects}
}
\value{
a sample details XML:xmlNode node
}
\description{
Create a sample details node assembled from one or more sample XML::xmlNode
See \url{http://genologics.com/files/permanent/API/latest/data_smp.html#element_details}
}

