% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Udf.R
\name{create_udf_node}
\alias{create_udf_node}
\title{Create a UDF XML::xmlNode and possibly add it to a parent node}
\usage{
create_udf_node(x, namespace = "udf", parent = NULL, ...)
}
\arguments{
\item{x}{a named character vector with at least "name", "type" and "unit" elements
but it may optionally include a "value" element.  If value is not provided then
the value in the returned node is empty.}

\item{namespace}{character, the xmlNamespace for the new xmlNode}

\item{parent}{optional xmlNode to whom to add the new node}

\item{...}{further arguments for \code{\link{newXMLNode}}}
}
\value{
udf XML::xmlNode
}
\description{
Suppose you have the following UDF attributes and value...
x <- c(unit = "uL", type = "Numeric", name = "facs.Padding volume", value = 0.6)
Below is the expected appearance of the output ...
<udf:field unit="uL" type="Numeric" name="facs.Padding volume">0.6</udf:field>
}

