% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file_transfer.R
\name{duck_download}
\alias{duck_download}
\title{Download a file using duck}
\usage{
duck_download(url, dest, username = "foo", password = "bar")
}
\arguments{
\item{url}{}

\item{dest}{character destination filename, by default the basename of the URL}

\item{username}{the username (required)}

\item{password}{the password (required)}
}
\value{
integer with 0 for success
}
\description{
Download a file using duck
}
\seealso{
Other file_transfer: \code{\link{duck_upload}};
  \code{\link{has_duck}}
}

