% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/file_transfer.R
\name{duck_upload}
\alias{duck_upload}
\title{Upload a file using duck}
\usage{
duck_upload(filename, url, username = "foo", password = "bar")
}
\arguments{
\item{filename}{the fully qualified filename of file to upload}

\item{url}{the destination url}

\item{username}{the username (required)}

\item{password}{the password (required)}
}
\value{
integer 0 for success
}
\description{
Upload a file using duck
}
\seealso{
Other file_transfer: \code{\link{duck_download}};
  \code{\link{has_duck}}; \code{\link{has_rsync}};
  \code{\link{has_scp}}; \code{\link{rsync_download}};
  \code{\link{rsync_upload}}; \code{\link{scp_download}};
  \code{\link{scp_upload}}
}

