% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Udf.R
\name{extract_udfs}
\alias{extract_udfs}
\title{Extract a named list of udf vectors from an xmlNode}
\usage{
extract_udfs(x)
}
\arguments{
\item{x}{XML::xmlNode possibly bearing udf fields}
}
\value{
a named list of lists where each sublist is comprised of
   \itemize{
      \item name character
      \item type character
      \item value data type varies
   }
 Or empty list if the xmlNode has no fields
}
\description{
Extract a named list of udf vectors from an xmlNode
}
\seealso{
Other Node Udf: \code{\link{NodeRefClass_get_udfs}};
  \code{\link{get_udfs}}; \code{\link{set_udfs}}
}

