% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Field.R
\docType{class}
\name{FieldRefClass-class}
\alias{FieldRefClass}
\alias{FieldRefClass-class}
\title{A User Defined Field configuration representation that subclasses from NodeRefClass}
\description{
A User Defined Field configuration representation that subclasses from NodeRefClass
}
\section{Fields}{

\describe{
\item{\code{name}}{character}

\item{\code{type}}{Field content type}

\item{\code{attach_to_name}}{character}

\item{\code{is_required}}{logical}

\item{\code{attach_to_category}}{character}

\item{\code{first_preset_is_default_value}}{logical}

\item{\code{preset}}{character (vector)}
}}
\seealso{
Other Node: \code{\link{ArtifactGroupRefClass}},
  \code{\link{ArtifactGroupRefClass-class}};
  \code{\link{ArtifactRefClass}},
  \code{\link{ArtifactRefClass-class}};
  \code{\link{ContainerRefClass}},
  \code{\link{ContainerRefClass-class}};
  \code{\link{ContainerTypeRefClass}},
  \code{\link{ContainerTypeRefClass-class}};
  \code{\link{ExceptionRefClass}},
  \code{\link{ExceptionRefClass-class}};
  \code{\link{FieldRefClass_show}};
  \code{\link{FileRefClass}},
  \code{\link{FileRefClass-class}};
  \code{\link{InputOutputMapRefClass}},
  \code{\link{InputOutputMapRefClass-class}};
  \code{\link{NodeRefClass_BROWSE}};
  \code{\link{NodeRefClass_DELETE}};
  \code{\link{NodeRefClass_GET}};
  \code{\link{NodeRefClass_POST}};
  \code{\link{NodeRefClass_PUT}};
  \code{\link{NodeRefClass_add_namespaces}};
  \code{\link{NodeRefClass_get_attached_files}};
  \code{\link{NodeRefClass_get_childv}};
  \code{\link{NodeRefClass_get_field}};
  \code{\link{NodeRefClass_get_limsid}};
  \code{\link{NodeRefClass_get_type}},
  \code{\link{NodeRefClass_get_type}};
  \code{\link{NodeRefClass_has_child}};
  \code{\link{NodeRefClass_has_lims}};
  \code{\link{NodeRefClass_new}};
  \code{\link{NodeRefClass_required_namespaces}};
  \code{\link{NodeRefClass_set_udfs}};
  \code{\link{NodeRefClass_show}};
  \code{\link{NodeRefClass_toString}},
  \code{\link{NodeRefClass_toString}};
  \code{\link{NodeRefClass_unames}};
  \code{\link{NodeRefClass_update}};
  \code{\link{NodeRefClass}},
  \code{\link{NodeRefClass-class}};
  \code{\link{ProcessRefClass}},
  \code{\link{ProcessRefClass-class}};
  \code{\link{ProcessTypeRefClass}},
  \code{\link{ProcessTypeRefClass-class}};
  \code{\link{ProjectRefClass}},
  \code{\link{ProjectRefClass-class}};
  \code{\link{ProjectsRefClass}},
  \code{\link{ProjectsRefClass-class}};
  \code{\link{ResearcherRefClass}},
  \code{\link{ResearcherRefClass-class}};
  \code{\link{SampleRefClass}},
  \code{\link{SampleRefClass-class}};
  \code{\link{udf_type2R}}
}

