% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\name{FieldRefClass_DELETE}
\alias{FieldRefClass_DELETE}
\title{DELETE is disallowed for field configuration}
\description{
DELETE is disallowed for field configuration
}
\seealso{
Other Field: \code{\link{FieldRefClass_POST}},
  \code{\link{FieldRefClass_get_type}}
}

