% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Field.R
\name{FieldRefClass_get_type}
\alias{FieldRefClass_get_type}
\title{Get the value of the Udf type}
\value{
character the type name or "" if none
}
\description{
Get the value of the Udf type
}
\seealso{
Other Field: \code{\link{FieldRefClass_DELETE}};
  \code{\link{FieldRefClass_POST}}
}

