% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\name{FieldRefClass_POST}
\alias{FieldRefClass_POST}
\title{POST is disallowed for field configuration}
\description{
POST is disallowed for field configuration
}
\seealso{
Other Field: \code{\link{FieldRefClass_DELETE}},
  \code{\link{FieldRefClass_get_type}}
}

