% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Node.R
\name{get_childvalue}
\alias{get_childvalue}
\title{Retrieve the value of a child node(s)}
\usage{
get_childvalue(x, name)
}
\arguments{
\item{x}{XML::xmlNode}

\item{name}{the name of the child(ren)}
}
\value{
named character vector of child values, possibly ""
}
\description{
Retrieve the value of a child node(s)
}

