% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_config}
\alias{get_config}
\title{Retrieve a configuration value}
\usage{
get_config(x, section, name, default = NULL)
}
\arguments{
\item{x}{the configuration list}

\item{section}{the name of the section}

\item{name}{the name of the tagged value, if missing then the section is returned}

\item{default}{the value to return if the tag doesn't exists (defaults to NULL)}
}
\value{
the tagged value or section requested or the 'default' value if not found
}
\description{
Retrieve a configuration value
}
