% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{get_NSMAP}
\alias{get_NSMAP}
\title{Retrieve the Genologics namespace map}
\usage{
get_NSMAP()
}
\value{
a named vector of xmlNameSpaces
}
\description{
Retrieve the Genologics namespace map
}
\seealso{
Other Lims: \code{\link{LimsRefClass_BROWSE}};
  \code{\link{LimsRefClass_DELETE}};
  \code{\link{LimsRefClass_POST}};
  \code{\link{LimsRefClass_PUSH}};
  \code{\link{LimsRefClass_PUT}};
  \code{\link{LimsRefClass_get_byLimsid}};
  \code{\link{LimsRefClass_uri}};
  \code{\link{LimsRefClass_validate_session}};
  \code{\link{LimsRefClass}},
  \code{\link{LimsRefClass-class}}
}

