% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Udf.R
\name{get_udfs}
\alias{get_udfs}
\title{Retrieve the values of one or more udfs}
\usage{
get_udfs(x, name)
}
\arguments{
\item{x}{XML::xmlNode object}

\item{name}{character one or more names}
}
\value{
named list of field values.  If no fields exists then NULL is return.
If a name is missing then
}
\description{
Retrieve the values of one or more udfs
}
\seealso{
Other Node Udf: \code{\link{NodeRefClass_get_udfs}};
  \code{\link{extract_udfs}}; \code{\link{set_udfs}}
}

