% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{get_uri}
\alias{get_uri}
\title{Get a uri with option to depaginate}
\usage{
get_uri(uri, lims, ..., depaginate = TRUE, verbose = FALSE, tries = 3)
}
\arguments{
\item{uri}{character, the uri to retrieve}

\item{lims}{a LimsRefClass node}

\item{...}{further arguments for httr::GET()}

\item{depaginate}{logical, if TRUE (the default) then depaginate the results}

\item{verbose}{logical if TRUE be verbose}

\item{tries}{numeric, allow up to this number of tries before failing}
}
\value{
XML::xmlNode
}
\description{
Get a uri with option to depaginate
}

