% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file_transfer.R
\name{has_duck}
\alias{has_duck}
\title{Test if duck is installed on the host platform
See \url{https://cyberduck.io/}}
\usage{
has_duck()
}
\value{
logical, TRUE if duck is present
}
\description{
Test if duck is installed on the host platform
See \url{https://cyberduck.io/}
}
\seealso{
Other file_transfer: \code{\link{duck_download}};
  \code{\link{duck_upload}}; \code{\link{has_scp}};
  \code{\link{scp_download}}; \code{\link{scp_upload}}
}

