% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file_transfer.R
\name{has_rsync}
\alias{has_rsync}
\title{Test if rsync is installed on the host platform
Perhaps silly but symmetric with others.}
\usage{
has_rsync()
}
\value{
logical, TRUE if rsync is present
}
\description{
Test if rsync is installed on the host platform
Perhaps silly but symmetric with others.
}
\seealso{
Other file_transfer: \code{\link{cp_download}};
  \code{\link{duck_download}}; \code{\link{duck_upload}};
  \code{\link{has_duck}}; \code{\link{has_scp}};
  \code{\link{rsync_download}}; \code{\link{rsync_upload}};
  \code{\link{scp_download}}; \code{\link{scp_upload}}
}

