% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_transfer.R
\name{has_scp}
\alias{has_scp}
\title{Test if scp is installed on the host platform
Perhaps silly but symmetric with others.}
\usage{
has_scp()
}
\value{
logical, TRUE if scp is present
}
\description{
Test if scp is installed on the host platform
Perhaps silly but symmetric with others.
}
\seealso{
Other file_transfer: \code{\link{cp_download}},
  \code{\link{duck_download}}, \code{\link{duck_upload}},
  \code{\link{has_duck}}, \code{\link{has_rsync}},
  \code{\link{make_remote_directory}},
  \code{\link{remote_directory_exists}},
  \code{\link{rsync_download}}, \code{\link{rsync_upload}},
  \code{\link{scp_download}}, \code{\link{scp_upload}}
}
