% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputOutputMap.R
\docType{class}
\name{InputOutputMapRefClass-class}
\alias{InputOutputMapRefClass}
\alias{InputOutputMapRefClass-class}
\title{An Input-Output-Map representation that subclasses from NodeRefClass}
\description{
An Input-Output-Map representation that subclasses from NodeRefClass
}
\section{Fields}{

\describe{
\item{\code{input_uri}}{character}

\item{\code{input_limsid}}{character}

\item{\code{post_process_uri}}{character}

\item{\code{output_uri}}{character}

\item{\code{output_limsid}}{character}

\item{\code{output_generation_type}}{character}

\item{\code{output_type}}{character}
}}
\seealso{
Other Node: \code{\link{ArtifactGroupRefClass-class}},
  \code{\link{ArtifactRefClass-class}},
  \code{\link{ContainerRefClass-class}},
  \code{\link{ContainerTypeRefClass-class}},
  \code{\link{ExceptionRefClass-class}},
  \code{\link{FieldRefClass-class}},
  \code{\link{FieldRefClass_show}},
  \code{\link{FileRefClass-class}},
  \code{\link{NodeRefClass-class}},
  \code{\link{NodeRefClass_ATTACH}},
  \code{\link{NodeRefClass_BROWSE}},
  \code{\link{NodeRefClass_DELETE}},
  \code{\link{NodeRefClass_GET}},
  \code{\link{NodeRefClass_POST}},
  \code{\link{NodeRefClass_PUT}},
  \code{\link{NodeRefClass_add_namespaces}},
  \code{\link{NodeRefClass_get_attached_files}},
  \code{\link{NodeRefClass_get_childv}},
  \code{\link{NodeRefClass_get_field}},
  \code{\link{NodeRefClass_get_limsid}},
  \code{\link{NodeRefClass_get_type}},
  \code{\link{NodeRefClass_has_child}},
  \code{\link{NodeRefClass_has_lims}},
  \code{\link{NodeRefClass_new}},
  \code{\link{NodeRefClass_required_namespaces}},
  \code{\link{NodeRefClass_set_udfs}},
  \code{\link{NodeRefClass_show}},
  \code{\link{NodeRefClass_toString}},
  \code{\link{NodeRefClass_unames}},
  \code{\link{NodeRefClass_update}},
  \code{\link{ProcessRefClass-class}},
  \code{\link{ProcessTypeRefClass-class}},
  \code{\link{ProjectRefClass-class}},
  \code{\link{ProjectsRefClass-class}},
  \code{\link{ResearcherRefClass-class}},
  \code{\link{SampleRefClass-class}},
  \code{\link{udf_type2R}}
}

