% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputOutputMap.R
\name{InputOutputMapRefClass_DELETE}
\alias{InputOutputMapRefClass_DELETE}
\title{DELETE is disallowed for artifacts}
\description{
DELETE is disallowed for artifacts
}
\seealso{
Other InputOutputMap: \code{\link{InputOutputMapRefClass_GET}},
  \code{\link{InputOutputMapRefClass_POST}},
  \code{\link{InputOutputMapRefClass_PUT}}
}
