% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputOutputMap.R
\name{InputOutputMapRefClass_GET}
\alias{InputOutputMapRefClass_GET}
\title{GET is disallowed for input-output-map}
\description{
GET is disallowed for input-output-map
}
\seealso{
Other InputOutputMap: \code{\link{InputOutputMapRefClass_DELETE}},
  \code{\link{InputOutputMapRefClass_POST}},
  \code{\link{InputOutputMapRefClass_PUT}}
}

