% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputOutputMap.R
\name{InputOutputMapRefClass_POST}
\alias{InputOutputMapRefClass_POST}
\title{POST is disallowed for input-output-map}
\description{
POST is disallowed for input-output-map
}
\seealso{
Other InputOutputMap: \code{\link{InputOutputMapRefClass_DELETE}},
  \code{\link{InputOutputMapRefClass_GET}},
  \code{\link{InputOutputMapRefClass_PUT}}
}

