% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/InputOutputMap.R
\name{InputOutputMapRefClass_PUT}
\alias{InputOutputMapRefClass_PUT}
\title{PUT is disallowed for input-output-map}
\description{
PUT is disallowed for input-output-map
}
\seealso{
Other InputOutputMap: \code{\link{InputOutputMapRefClass_DELETE}};
  \code{\link{InputOutputMapRefClass_GET}};
  \code{\link{InputOutputMapRefClass_POST}}
}

