% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Instrument.R
\name{InstrumentRefClass_POST}
\alias{InstrumentRefClass_POST}
\title{POST is disallowed for instruments}
\description{
POST is disallowed for instruments
}
\seealso{
Other Instrument: \code{\link{InstrumentRefClass_PUT}}
}

