% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument.R
\name{InstrumentRefClass_PUT}
\alias{InstrumentRefClass_PUT}
\title{PUT is disallowed for instruments}
\description{
PUT is disallowed for instruments
}
\seealso{
Other Instrument: \code{\link{InstrumentRefClass_POST}}
}
