% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab.R
\docType{class}
\name{LabRefClass-class}
\alias{LabRefClass}
\alias{LabRefClass-class}
\title{A Lab/Account representation that subclasses from NodeRefClass}
\description{
A Lab/Account representation that subclasses from NodeRefClass
}
\section{Fields}{

\describe{
\item{\code{name}}{character}

\item{\code{website}}{character}

\item{\code{shipping_address}}{character vector}

\item{\code{billing_address}}{character vector}
}}

