% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\docType{class}
\name{LimsRefClass-class}
\alias{LimsRefClass}
\alias{LimsRefClass-class}
\title{Reference class for Lims instance}
\description{
Reference class for Lims instance
}
\section{Fields}{

\describe{
\item{\code{version}}{character Genologics LIMS version}

\item{\code{baseuri}}{character the base uri}

\item{\code{auth}}{httr::authenticate object for LIMS}

\item{\code{fileauth}}{httr::authenticate object for filestore}

\item{\code{handle}}{httr::handle object}
}}
\seealso{
Other Lims: \code{\link{LimsRefClass_BROWSE}};
  \code{\link{LimsRefClass_DELETE}};
  \code{\link{LimsRefClass_POST}};
  \code{\link{LimsRefClass_PUSH}};
  \code{\link{LimsRefClass_PUT}};
  \code{\link{LimsRefClass_get_byLimsid}};
  \code{\link{LimsRefClass_uri}};
  \code{\link{LimsRefClass_validate_session}};
  \code{\link{get_NSMAP}}
}

