% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_batchcreate}
\alias{LimsRefClass_batchcreate}
\title{Create one or more Nodes (samples, containers only)}
\arguments{
\item{x}{a list of one or more XML::xmlNode or NodeRefClass}

\item{asNode}{logical, if TRUE return NodeRefClass objects otherwise XML::xmlNode}

\item{...}{further arguments for httr::POST}
}
\value{
a list of NodeRefClass or NULL
}
\description{
Create one or more Nodes (samples, containers only)
}
\seealso{
Other Lims Node: \code{\link{LimsRefClass_batchretrieve}},
  \code{\link{LimsRefClass_batchupdate}},
  \code{\link{parse_node}}
}

