% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_batchretrieve}
\alias{LimsRefClass_batchretrieve}
\title{Get one or more nodes by uri by batch (artifacts, files, samples, containers only)}
\arguments{
\item{uri}{a vector of one or more uri for atomic entities in the GLS API}

\item{rel}{the relative name space into the "batch/retrieve"}

\item{asNode}{logical, if TRUE parse to the appropriate node type}

\item{rm_dups}{logical, if TRUE then remove duplicates}

\item{...}{further arguments for \code{batch_retrieve}}
}
\value{
a list of XML::xmlNode or NodeRefClass objects
}
\description{
Get one or more nodes by uri by batch (artifacts, files, samples, containers only)
}
\seealso{
Other Lims Node: \code{\link{LimsRefClass_batchcreate}};
  \code{\link{LimsRefClass_batchupdate}};
  \code{\link{parse_node}}
}

