% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_batchupdate}
\alias{LimsRefClass_batchupdate}
\title{Update one or more Nodes (artifacts, samples, containers only)}
\arguments{
\item{x}{a list of one or mode XML::xmlNode of NodeRefClass}

\item{asNode}{logical, if TRUE return NodeRefClass objects otherwise XML::xmlNode}

\item{...}{further arguments for httr::POST}
}
\value{
a list of NodeRefClass or NULL
}
\description{
Return order is enforced to be the same as the input order
}
\seealso{
Other Lims Node: \code{\link{LimsRefClass_batchcreate}},
  \code{\link{LimsRefClass_batchretrieve}},
  \code{\link{parse_node}}
}
