% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_BROWSE}
\alias{LimsRefClass_BROWSE}
\title{BROWSE a URI in a browser if in interactive session}
\arguments{
\item{x}{XML::xmlNode or NodeRefClass}

\item{...}{further arguments for httr::BROWSE}
}
\description{
BROWSE a URI in a browser if in interactive session
}
\seealso{
Other Lims: \code{\link{LimsRefClass_DELETE}};
  \code{\link{LimsRefClass_POST}};
  \code{\link{LimsRefClass_PUSH}};
  \code{\link{LimsRefClass_PUT}};
  \code{\link{LimsRefClass_check}};
  \code{\link{LimsRefClass_get_byLimsid}};
  \code{\link{LimsRefClass_uri}};
  \code{\link{LimsRefClass_userpwd}};
  \code{\link{LimsRefClass_validate_session}};
  \code{\link{LimsRefClass}},
  \code{\link{LimsRefClass-class}}; \code{\link{get_NSMAP}}
}

