% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_check}
\alias{LimsRefClass_check}
\title{Verify a response and return xmlNode, possible an exception node}
\arguments{
\item{rsp}{httr::response object}

\item{msg}{if an exception is encountered, attach this message (unless NULL)}
}
\value{
XML::xmlNode
}
\description{
Verify a response and return xmlNode, possible an exception node
}

