% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_create_exception}
\alias{LimsRefClass_create_exception}
\title{Create an exception node}
\arguments{
\item{message}{chracter, some error message}
}
\value{
XML::xmlNode
}
\description{
Create an exception node
}

