% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_DELETE}
\alias{LimsRefClass_DELETE}
\title{DELETE a resource}
\arguments{
\item{x}{NodeRefClass, XML::xmlNode or character uri to DELETE, generally a file node}

\item{...}{further arguments for httr::DELETE}
}
\value{
logical
}
\description{
Typically this is a file (resource = 'files')
}
\seealso{
Other Lims: \code{\link{LimsRefClass-class}},
  \code{\link{LimsRefClass_ATTACH}},
  \code{\link{LimsRefClass_BROWSE}},
  \code{\link{LimsRefClass_POST}},
  \code{\link{LimsRefClass_PUSH}},
  \code{\link{LimsRefClass_PUT}},
  \code{\link{LimsRefClass_check}},
  \code{\link{LimsRefClass_get_artifactgroups}},
  \code{\link{LimsRefClass_get_byLimsid}},
  \code{\link{LimsRefClass_uri}},
  \code{\link{LimsRefClass_userpwd}},
  \code{\link{LimsRefClass_validate_session}},
  \code{\link{get_NSMAP}}
}

