% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_GET}
\alias{LimsRefClass_GET}
\title{GET a resource, a wrapper around get_uri}
\arguments{
\item{uri}{character the uri to get}

\item{...}{further arguments for httr::GET including \code{query} list}

\item{depaginate}{logical, if TRUE then pool paginated nodes into one}

\item{asNode}{logical, if TRUE return a class that inherits NodeRefClass}
}
\value{
XML::xmlNode - possibly an error node
}
\description{
GET a resource, a wrapper around get_uri
}

