% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_artifactgroups}
\alias{LimsRefClass_get_artifactgroups}
\title{Get artifact group(s) in the system}
\arguments{
\item{artifactgroup}{a character vector of one or more artifact group names}
}
\value{
a named list of ArtifactGroupTypeRefClass objects or NULL
}
\description{
Get artifact group(s) in the system
}
\seealso{
Other Lims: \code{\link{LimsRefClass_ATTACH}};
  \code{\link{LimsRefClass_BROWSE}};
  \code{\link{LimsRefClass_DELETE}};
  \code{\link{LimsRefClass_POST}};
  \code{\link{LimsRefClass_PUSH}};
  \code{\link{LimsRefClass_PUT}};
  \code{\link{LimsRefClass_check}};
  \code{\link{LimsRefClass_get_byLimsid}};
  \code{\link{LimsRefClass_uri}};
  \code{\link{LimsRefClass_userpwd}};
  \code{\link{LimsRefClass_validate_session}};
  \code{\link{LimsRefClass}},
  \code{\link{LimsRefClass-class}}; \code{\link{get_NSMAP}}
}

