% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_artifacts}
\alias{LimsRefClass_get_artifacts}
\title{Get one or more artifacts using queries on name, type, process-type, working-flag
qc-flag, sample-name, samplelimsid, containername, containerlimsid, reagent-label}
\arguments{
\item{name}{one or more artifact names or NULL to ignore}

\item{type}{character one or more character types or NULL to ignore}

\item{process_type}{character parent process type or NULL to ignore}

\item{working_flag}{character 'true' or 'false'  or NULL to ignore}

\item{qc_flag}{character on of UNKNOWN, PASSED, FAILED, CONTINUE or NULL to ignore}

\item{sample_name}{character one or more submitted sample names or NULL to ignore}

\item{samplelimsid}{character one or more submitted sample limsid  or NULL to ignore}

\item{artifactgroup}{character one or more experiment names  or NULL to ignore}

\item{container_name}{character one or more container names or NULL to ignore}

\item{containerlimsid}{character one or more container limsid  or NULL to ignore}

\item{reagent-label}{character one or more reagent names or NULL to ignore}
}
\value{
a list of ArtifactRefClass objects or NULL
}
\description{
Get one or more artifacts using queries on name, type, process-type, working-flag
qc-flag, sample-name, samplelimsid, containername, containerlimsid, reagent-label
}

