% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_byLimsid}
\alias{LimsRefClass_get_byLimsid}
\title{Retrieve a resource by limsid}
\arguments{
\item{lismid}{character, one or more limsids}

\item{resource}{character, one resource to search, by default 'artifacts'}
}
\value{
a list of NodeRefClass objects
}
\description{
Retrieve a resource by limsid
}
\seealso{
Other Lims: \code{\link{LimsRefClass-class}},
  \code{\link{LimsRefClass_ATTACH}},
  \code{\link{LimsRefClass_BROWSE}},
  \code{\link{LimsRefClass_DELETE}},
  \code{\link{LimsRefClass_POST}},
  \code{\link{LimsRefClass_PUSH}},
  \code{\link{LimsRefClass_PUT}},
  \code{\link{LimsRefClass_check}},
  \code{\link{LimsRefClass_get_artifactgroups}},
  \code{\link{LimsRefClass_uri}},
  \code{\link{LimsRefClass_userpwd}},
  \code{\link{LimsRefClass_validate_session}},
  \code{\link{get_NSMAP}}
}
