% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_containers}
\alias{LimsRefClass_get_containers}
\title{Get one or more containers by name, state, etc}
\arguments{
\item{optional}{name a character vector of one or more names}

\item{optional}{type character of one or more container types ("384 well plate", etc)}

\item{optional}{state character of one or more contain states ("Discarded", "Populated",...)}

\item{optional}{last_modified a character vector of last modification date in YYYY-MM-DDThh:mm:ssTZD format}
}
\value{
a named list of ContainerRefClass or NULL
}
\description{
Get one or more containers by name, state, etc
}
\seealso{
Other Lims Container: \code{\link{LimsRefClass_get_containertypes}},
  \code{\link{create_container_node}}
}
