% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_containertypes}
\alias{LimsRefClass_get_containertypes}
\title{Get the container type(s) in the system}
\arguments{
\item{optional}{name a character vector of one or more container type names}
}
\value{
a named list of ContainerTypeRefClass objects or NULL
}
\description{
Get the container type(s) in the system
}
\seealso{
Other Lims Container: \code{\link{LimsRefClass_get_containers}};
  \code{\link{create_container_node}}
}

