% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_fields}
\alias{LimsRefClass_get_fields}
\title{Get one or more fields (UDF) as Nodes}
\arguments{
\item{name}{one or more optional UDF names}

\item{attach_to_name}{one name of a container, sample, project, process}

\item{attach-to-category}{If 'attach_to_name' is the name of a process, 
specify 'ProcessType'. Must not be provided otherwise.}

\item{...}{further arguments for GET method}
}
\value{
a list of FieldRefClass or NULL
}
\description{
Get one or more fields (UDF) as Nodes
}
