% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_instruments}
\alias{LimsRefClass_get_instruments}
\title{Retrieve a list of InstrumentRefClass or data.frame of the good stuff}
\arguments{
\item{optional}{name character a vector of one or more names}

\item{asDataFrame}{logical, return a data.frame or list of Nodes}
}
\value{
a list of  InstrumentRefClass, a data frame or NULL
}
\description{
Retrieve a list of InstrumentRefClass or data.frame of the good stuff
}

