% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_labs}
\alias{LimsRefClass_get_labs}
\title{Get one or more Labs - does not leverage /batch/retrieve resources
but provides similar behavior.}
\arguments{
\item{name}{optional lab name}

\item{last_modified}{optional character vector in YYYY-MM-DDThh:mm:ssTZD format}
}
\value{
a list of LabRefClass or NULL
}
\description{
Get one or more Labs - does not leverage /batch/retrieve resources
but provides similar behavior.
}
