% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_max_requests}
\alias{LimsRefClass_get_max_requests}
\title{Retrieve the max requests vakue by name}
\arguments{
\item{name}{character the namespace to retrieve singualr or plural forms
of artifacts, samples, containers or files}
}
\value{
numeric, the max requests per batch call
}
\description{
Retrieve the max requests vakue by name
}

