% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_or_create_containers}
\alias{LimsRefClass_get_or_create_containers}
\title{Get or create containers by name and type}
\arguments{
\item{name}{character one or more container names}

\item{ctype}{character one or more container types - only one type is allowed}
}
\value{
a list of one or more ContainerRefClass or NULL
}
\description{
Get or create containers by name and type
}
