% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_processes}
\alias{LimsRefClass_get_processes}
\title{Get one or more Processes - does not leverage /batch/retrieve resources
but provides similar behavior.}
\arguments{
\item{last_modified}{optional character vector in YYYY-MM-DDThh:mm:ssTZD format}

\item{type}{optional character of the process type}

\item{inputartifactlimsid}{optional character of an input artifact limsid}

\item{technamefirst}{optional technician's first name}

\item{technamelast}{optional technician's last name}

\item{projectname}{optional project name}
}
\value{
a list of ProcessRefClass or NULL
}
\description{
Get one or more Processes - does not leverage /batch/retrieve resources
but provides similar behavior.
}
\seealso{
Other Lims Process: \code{\link{LimsRefClass_get_processtypes}}
}

