% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_processtypes}
\alias{LimsRefClass_get_processtypes}
\title{Get one or more process-types as Nodes}
\arguments{
\item{displayname}{optional project name}
}
\value{
a list of NodeRefClass or NULL
}
\description{
Get one or more process-types as Nodes
}
\seealso{
Other Lims Process: \code{\link{LimsRefClass_get_processes}}
}
