% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_projects}
\alias{LimsRefClass_get_projects}
\title{Get one or more Projects - does not leverage /batch/retrieve resources
but provides similar behavior.}
\arguments{
\item{name}{optional project name}

\item{last_modified}{optional character vector in YYYY-MM-DDThh:mm:ssTZD format}
}
\value{
a list of ProjectRefClass or NULL
}
\description{
Get one or more Projects - does not leverage /batch/retrieve resources
but provides similar behavior.
}

