% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_researchers}
\alias{LimsRefClass_get_researchers}
\title{Retrieve a list of ResearcherRefNodes or a data.frame of the good stuff}
\arguments{
\item{optional}{username character a vector of one or more user names like 'btupper' etc.}
}
\value{
a list of  ResearcherRefClass, a data frame or NULL
}
\description{
Retrieve a list of ResearcherRefNodes or a data.frame of the good stuff
}
\seealso{
Other Lims Researcher: \code{\link{create_researcher_node}}
}

