% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_samples}
\alias{LimsRefClass_get_samples}
\title{Get one or more samples using queries on name, projectlimsid, projectname
It is possible to also filter the query on UDF values but it may be easier to}
\arguments{
\item{optional}{name a character vector of one or more names}

\item{optional}{projectlimsid character of one or more projectlimsid values}

\item{optional}{projectname character of one or more projectname values}
}
\value{
a named list of container XML::xmlNode
}
\description{
Get one or more samples using queries on name, projectlimsid, projectname
It is possible to also filter the query on UDF values but it may be easier to
}
\seealso{
Other Lims Sample: \code{\link{create_sample_node}}
}

