% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_get_workflows}
\alias{LimsRefClass_get_workflows}
\title{Retrieve a list of WorkflowRefClass}
\arguments{
\item{name}{character vector of one or more workflow names}

\item{form}{character of 'uri' or 'Node'}
}
\value{
character vector or list with zero or more uri/WorkflowRefClass or NULL
}
\description{
Retrieve a list of WorkflowRefClass
}
