% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_LIST}
\alias{LimsRefClass_LIST}
\title{LIST the URI's a resource with qualifiers}
\arguments{
\item{resource}{character the uri to get}

\item{n}{numeric, the maximum number of URI, NA to get all}

\item{...}{further arguments for httr::GET including \code{query} list}
}
\value{
character vector of zero or more URI
}
\description{
LIST the URI's a resource with qualifiers
}
\examples{
\dontrun{
    # list the samples in a project
    ss <- lims$LIST('samples', projectname = 'foobar')
}
}
