% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_name_host}
\alias{LimsRefClass_name_host}
\title{Modify the URL provided to point to the named server}
\arguments{
\item{x}{a character URL like "http://localhost:9080/api/v2/processes/24-2792440"}
}
\value{
an updated URL the explicitly names the host
}
\description{
Modify the URL provided to point to the named server
}
