% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_PUSH}
\alias{LimsRefClass_PUSH}
\title{PUSH a file - not really a RESTfule action but a combination of steps}
\arguments{
\item{x}{XML::xmlNode of the artifact to attach to}

\item{...}{further arguments for httr::GET/DELETE/POST}

\item{filename}{character, the fully qualified name of the file we are pushing}
}
\value{
XML::xmlNode or FileRefClass
}
\description{
given an artifact node and a filename
if the artifact has a file then
   DELETE the file resource
create an unresolved file resource
POST the unresolved file resource to 'glsstore' to get a resolved file resource
Upload the file (scp, cp, or curl)
POST the resolved file resource to 'files'
return the resolved file resource
}
\seealso{
Other Lims: \code{\link{LimsRefClass_BROWSE}};
  \code{\link{LimsRefClass_DELETE}};
  \code{\link{LimsRefClass_POST}};
  \code{\link{LimsRefClass_PUT}};
  \code{\link{LimsRefClass_get_byLimsid}};
  \code{\link{LimsRefClass_uri}};
  \code{\link{LimsRefClass_validate_session}};
  \code{\link{LimsRefClass}},
  \code{\link{LimsRefClass-class}}; \code{\link{get_NSMAP}}
}

