% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_PUT}
\alias{LimsRefClass_PUT}
\title{PUT a resource}
\arguments{
\item{x}{xmlNode to put}

\item{...}{further arguments for httr::PUT}
}
\value{
xmlNode
}
\description{
PUT a resource
}
\seealso{
Other Lims: \code{\link{LimsRefClass_BROWSE}};
  \code{\link{LimsRefClass_DELETE}};
  \code{\link{LimsRefClass_POST}};
  \code{\link{LimsRefClass_PUSH}};
  \code{\link{LimsRefClass_check}};
  \code{\link{LimsRefClass_get_artifactgroups}};
  \code{\link{LimsRefClass_get_byLimsid}};
  \code{\link{LimsRefClass_uri}};
  \code{\link{LimsRefClass_userpwd}};
  \code{\link{LimsRefClass_validate_session}};
  \code{\link{LimsRefClass}},
  \code{\link{LimsRefClass-class}}; \code{\link{get_NSMAP}}
}

