% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_set_max_requests}
\alias{LimsRefClass_set_max_requests}
\title{Set the max requests value by name}
\arguments{
\item{value}{numeric named vector of values to set.  Here are the defaults
\itemize{
    \item{artifacts = 200}
    \item{containers = 50}
    \item{samples = 400}
    \item{files = 100}
}}
}
\value{
numeric, the max requests per batch call
}
\description{
Set the max requests value by name
}
