% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_show Sample}
\alias{LimsRefClass_show Sample}
\title{Print a pretty summary}
\arguments{
\item{prefix}{character - perhaps number of spaces to prefix the output}
}
\description{
Print a pretty summary
}

