% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_uri}
\alias{LimsRefClass_uri}
\title{Build a uri staring with the baseuri}
\arguments{
\item{...}{one or more character segments to append to the base}

\item{base}{character, the base uri, if NULL then use this object's baseuri property}
}
\value{
character uri
}
\description{
Build a uri staring with the baseuri
}
\examples{
\dontrun{
   my_uri <- ref$uri("containers")
}
}
\seealso{
Other Lims: \code{\link{LimsRefClass-class}},
  \code{\link{LimsRefClass_BROWSE}},
  \code{\link{LimsRefClass_DELETE}},
  \code{\link{LimsRefClass_POST}},
  \code{\link{LimsRefClass_PUSH}},
  \code{\link{LimsRefClass_PUT}},
  \code{\link{LimsRefClass_check}},
  \code{\link{LimsRefClass_get_artifactgroups}},
  \code{\link{LimsRefClass_get_byLimsid}},
  \code{\link{LimsRefClass_userpwd}},
  \code{\link{LimsRefClass_validate_session}},
  \code{\link{get_NSMAP}}
}

