% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_userpwd}
\alias{LimsRefClass_userpwd}
\title{Retrieve the user and password}
\arguments{
\item{what}{character to specify which authorization ('lims' or 'file')
by default lims is returned}
}
\value{
character vector of [username, password] or NULL
}
\description{
Retrieve the user and password
}
\seealso{
Other Lims: \code{\link{LimsRefClass_ATTACH}};
  \code{\link{LimsRefClass_BROWSE}};
  \code{\link{LimsRefClass_DELETE}};
  \code{\link{LimsRefClass_POST}};
  \code{\link{LimsRefClass_PUSH}};
  \code{\link{LimsRefClass_PUT}};
  \code{\link{LimsRefClass_check}};
  \code{\link{LimsRefClass_get_artifactgroups}};
  \code{\link{LimsRefClass_get_byLimsid}};
  \code{\link{LimsRefClass_uri}};
  \code{\link{LimsRefClass_validate_session}};
  \code{\link{LimsRefClass}},
  \code{\link{LimsRefClass-class}}; \code{\link{get_NSMAP}}
}

