% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_validate_session}
\alias{LimsRefClass_validate_session}
\title{Validate the session by testing the version}
\value{
logical, TRUE if OK
}
\description{
Validate the session by testing the version
}
\seealso{
Other Lims: \code{\link{LimsRefClass_BROWSE}};
  \code{\link{LimsRefClass_DELETE}};
  \code{\link{LimsRefClass_POST}};
  \code{\link{LimsRefClass_PUSH}};
  \code{\link{LimsRefClass_PUT}};
  \code{\link{LimsRefClass_check}};
  \code{\link{LimsRefClass_get_artifactgroups}};
  \code{\link{LimsRefClass_get_byLimsid}};
  \code{\link{LimsRefClass_uri}};
  \code{\link{LimsRefClass_userpwd}};
  \code{\link{LimsRefClass}},
  \code{\link{LimsRefClass-class}}; \code{\link{get_NSMAP}}
}

