% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{list_resource}
\alias{list_resource}
\title{List URIs in a resource such as samples or containers.}
\usage{
list_resource(lims, resource, n = NA, ...)
}
\arguments{
\item{lims}{the LimsRefClass object to query}

\item{resource}{character the uri to get}

\item{n}{numeric, the maximum number of URI, NA to get all}

\item{...}{further arguments for httr::GET including \code{query} list}
}
\value{
character vector of zero or more URI
}
\description{
List URIs in a resource such as samples or containers.
}
\examples{
\dontrun{
    # list the samples in a project
    ss <- list_resources(lims,'samples', projectname = 'foobar')
}
}
