% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_transfer.R
\name{make_remote_directory}
\alias{make_remote_directory}
\title{Make a remote directory specified by a URL using 
 \code{cp}, \code{scp} or \code{curl}.}
\usage{
make_remote_directory(uri, username = "foo", password = "bar",
  use = c("cp", "scp", "curl")[2], verbose = FALSE)
}
\arguments{
\item{uri}{character}

\item{username}{the username (required)}

\item{password}{the password (required for curl)}

\item{use}{character one of 'cp', 'scp' or 'curl'}

\item{verbose}{logical, if TRUE then echo the command string}
}
\value{
integer with 0 for success
}
\description{
\code{curl} builds paths as needed (recursively) so if you select
\code{curl} we return success without doing anything.
}
\seealso{
Other file_transfer: \code{\link{cp_download}},
  \code{\link{duck_download}}, \code{\link{duck_upload}},
  \code{\link{has_duck}}, \code{\link{has_rsync}},
  \code{\link{has_scp}},
  \code{\link{remote_directory_exists}},
  \code{\link{rsync_download}}, \code{\link{rsync_upload}},
  \code{\link{scp_download}}, \code{\link{scp_upload}}
}
