% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Node.R
\docType{class}
\name{NodeRefClass-class}
\alias{NodeRefClass}
\alias{NodeRefClass-class}
\title{The root class that wraps XML::xmlNode}
\description{
The root class that wraps XML::xmlNode
}
\section{Fields}{

\describe{
\item{\code{lims}}{a LimsRefClass object (possibly NULL)}

\item{\code{uri}}{character the uri of the node (possibly "")}

\item{\code{limsid}}{character the limsid of the node (possibly "")}

\item{\code{ns}}{character the XML::xmlNamespace object (possibly NULL)}

\item{\code{node}}{XML::xmlNode external pointer to the representation (possibly NULL)}
}}
\seealso{
Other Node: \code{\link{ArtifactRefClass}},
  \code{\link{ArtifactRefClass-class}};
  \code{\link{ContainerRefClass}},
  \code{\link{ContainerRefClass-class}};
  \code{\link{FileRefClass}},
  \code{\link{FileRefClass-class}};
  \code{\link{InputOutputMapRefClass}},
  \code{\link{InputOutputMapRefClass-class}};
  \code{\link{NodeRefClass_DELETE}};
  \code{\link{NodeRefClass_GET}};
  \code{\link{NodeRefClass_POST}};
  \code{\link{NodeRefClass_PUT}};
  \code{\link{NodeRefClass_get_field}};
  \code{\link{NodeRefClass_get_type}},
  \code{\link{NodeRefClass_get_type}};
  \code{\link{NodeRefClass_has_child}};
  \code{\link{NodeRefClass_has_lims}};
  \code{\link{NodeRefClass_new}};
  \code{\link{NodeRefClass_show}};
  \code{\link{NodeRefClass_toString}};
  \code{\link{NodeRefClass_unames get the unique child
  names}}; \code{\link{NodeRefClass_update}};
  \code{\link{ProcessRefClass}},
  \code{\link{ProcessRefClass-class}};
  \code{\link{ResearcherRefClass}},
  \code{\link{ResearcherRefClass-class}};
  \code{\link{SampleRefClass}},
  \code{\link{SampleRefClass-class}};
  \code{\link{udf_type2R}}
}

