% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Node.R
\name{NodeRefClass_get_udfs}
\alias{NodeRefClass_get_udfs}
\title{Extract the UDFs into a list of udf vectors each with 'name', 'type' and 'value' elements}
\value{
a named list of lists where each sublist is comprised of
   \itemize{
      \item name character
      \item type character
      \item value data type varies
   }
 Or empty list if the xmlNode has no fields
}
\description{
Extract the UDFs into a list of udf vectors each with 'name', 'type' and 'value' elements
}
\seealso{
Other Node Udf: \code{\link{extract_udfs}};
  \code{\link{get_udfs}}
}

