% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Node.R
\name{NodeRefClass_has_child}
\alias{NodeRefClass_has_child}
\title{Determine if this node has a child, test by name of XML::xmlNode reference}
\arguments{
\item{x}{one or more names to test or one or more XML::xmlNode objects}
}
\value{
logical vector, named if \code{x} is character
}
\description{
Determine if this node has a child, test by name of XML::xmlNode reference
}
\seealso{
Other Node: \code{\link{ArtifactRefClass}},
  \code{\link{ArtifactRefClass-class}};
  \code{\link{ContainerRefClass}},
  \code{\link{ContainerRefClass-class}};
  \code{\link{FieldRefClass_show}};
  \code{\link{FieldRefClass}},
  \code{\link{FieldRefClass-class}};
  \code{\link{FileRefClass}},
  \code{\link{FileRefClass-class}};
  \code{\link{InputOutputMapRefClass}},
  \code{\link{InputOutputMapRefClass-class}};
  \code{\link{NodeRefClass_DELETE}};
  \code{\link{NodeRefClass_GET}};
  \code{\link{NodeRefClass_POST}};
  \code{\link{NodeRefClass_PUT}};
  \code{\link{NodeRefClass_get_childv}};
  \code{\link{NodeRefClass_get_field}};
  \code{\link{NodeRefClass_get_type}},
  \code{\link{NodeRefClass_get_type}};
  \code{\link{NodeRefClass_has_lims}};
  \code{\link{NodeRefClass_new}};
  \code{\link{NodeRefClass_show}};
  \code{\link{NodeRefClass_toString}};
  \code{\link{NodeRefClass_unames}};
  \code{\link{NodeRefClass_update}};
  \code{\link{NodeRefClass}},
  \code{\link{NodeRefClass-class}};
  \code{\link{ProcessRefClass}},
  \code{\link{ProcessRefClass-class}};
  \code{\link{ResearcherRefClass}},
  \code{\link{ResearcherRefClass-class}};
  \code{\link{SampleRefClass}},
  \code{\link{SampleRefClass-class}};
  \code{\link{udf_type2R}}
}

