% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Node.R
\name{NodeRefClass_new}
\alias{NodeRefClass_new}
\title{Called when the object is instantiated.  NodeRefClass is a convenience wrapper
for XML::xmlNode resources used to work with the GLS API.}
\arguments{
\item{node}{either XML::xmlNode or a uri that points to such}

\item{lims}{LimsRefClass instance}
}
\value{
an instance of NodeRefClass
}
\description{
Called when the object is instantiated.  NodeRefClass is a convenience wrapper
for XML::xmlNode resources used to work with the GLS API.
}
\seealso{
Other Node: \code{\link{ArtifactGroupRefClass-class}},
  \code{\link{ArtifactRefClass-class}},
  \code{\link{ContainerRefClass-class}},
  \code{\link{ContainerTypeRefClass-class}},
  \code{\link{ExceptionRefClass-class}},
  \code{\link{FieldRefClass-class}},
  \code{\link{FieldRefClass_show}},
  \code{\link{FileRefClass-class}},
  \code{\link{InputOutputMapRefClass-class}},
  \code{\link{NodeRefClass-class}},
  \code{\link{NodeRefClass_BROWSE}},
  \code{\link{NodeRefClass_DELETE}},
  \code{\link{NodeRefClass_GET}},
  \code{\link{NodeRefClass_POST}},
  \code{\link{NodeRefClass_PUT}},
  \code{\link{NodeRefClass_add_namespaces}},
  \code{\link{NodeRefClass_get_childv}},
  \code{\link{NodeRefClass_get_field}},
  \code{\link{NodeRefClass_get_type}},
  \code{\link{NodeRefClass_has_child}},
  \code{\link{NodeRefClass_has_lims}},
  \code{\link{NodeRefClass_required_namespaces}},
  \code{\link{NodeRefClass_set_udfs}},
  \code{\link{NodeRefClass_show}},
  \code{\link{NodeRefClass_toString}},
  \code{\link{NodeRefClass_unames}},
  \code{\link{NodeRefClass_update}},
  \code{\link{ProcessRefClass-class}},
  \code{\link{ProcessTypeRefClass-class}},
  \code{\link{ProjectRefClass-class}},
  \code{\link{ProjectsRefClass-class}},
  \code{\link{ResearcherRefClass-class}},
  \code{\link{SampleRefClass-class}},
  \code{\link{udf_type2R}}
}

