% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Node.R
\name{NodeRefClass_set_udfs}
\alias{NodeRefClass_set_udfs}
\title{Set the value of the value of  one or more UDF children.  If the
field does not exist it is created and added, otherwise it is updated.}
\arguments{
\item{v}{a list of one or more udf vectors
each udf vector must have \code{name}, \code{type} and \code{value}}
}
\value{
a reference to this (updated) NodeRefClass object
}
\description{
Set the value of the value of  one or more UDF children.  If the
field does not exist it is created and added, otherwise it is updated.
}
\seealso{
Other Node: \code{\link{ArtifactGroupRefClass}},
  \code{\link{ArtifactGroupRefClass-class}};
  \code{\link{ArtifactRefClass}},
  \code{\link{ArtifactRefClass-class}};
  \code{\link{ContainerRefClass}},
  \code{\link{ContainerRefClass-class}};
  \code{\link{ContainerTypeRefClass}},
  \code{\link{ContainerTypeRefClass-class}};
  \code{\link{ExceptionRefClass}},
  \code{\link{ExceptionRefClass-class}};
  \code{\link{FieldRefClass_show}};
  \code{\link{FieldRefClass}},
  \code{\link{FieldRefClass-class}};
  \code{\link{FileRefClass}},
  \code{\link{FileRefClass-class}};
  \code{\link{InputOutputMapRefClass}},
  \code{\link{InputOutputMapRefClass-class}};
  \code{\link{NodeRefClass_BROWSE}};
  \code{\link{NodeRefClass_DELETE}};
  \code{\link{NodeRefClass_GET}};
  \code{\link{NodeRefClass_POST}};
  \code{\link{NodeRefClass_PUT}};
  \code{\link{NodeRefClass_add_namespaces}};
  \code{\link{NodeRefClass_get_childv}};
  \code{\link{NodeRefClass_get_field}};
  \code{\link{NodeRefClass_get_limsid}};
  \code{\link{NodeRefClass_get_type}},
  \code{\link{NodeRefClass_get_type}};
  \code{\link{NodeRefClass_has_child}};
  \code{\link{NodeRefClass_has_lims}};
  \code{\link{NodeRefClass_new}};
  \code{\link{NodeRefClass_required_namespaces}};
  \code{\link{NodeRefClass_show}};
  \code{\link{NodeRefClass_toString}},
  \code{\link{NodeRefClass_toString}};
  \code{\link{NodeRefClass_unames}};
  \code{\link{NodeRefClass_update}};
  \code{\link{NodeRefClass}},
  \code{\link{NodeRefClass-class}};
  \code{\link{ProcessRefClass}},
  \code{\link{ProcessRefClass-class}};
  \code{\link{ProcessTypeRefClass}},
  \code{\link{ProcessTypeRefClass-class}};
  \code{\link{ProjectRefClass}},
  \code{\link{ProjectRefClass-class}};
  \code{\link{ProjectsRefClass}},
  \code{\link{ProjectsRefClass-class}};
  \code{\link{ResearcherRefClass}},
  \code{\link{ResearcherRefClass-class}};
  \code{\link{SampleRefClass}},
  \code{\link{SampleRefClass-class}};
  \code{\link{udf_type2R}}
}

