% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{parse_node}
\alias{parse_node}
\title{Update a batch of XML::xmlNodes
Convert a node to an object inheriting from NodeRefClass}
\usage{
parse_node(node, lims)
}
\arguments{
\item{node}{XML::xmlNode}

\item{lims}{LimsRefClass object}
}
\description{
Update a batch of XML::xmlNodes
Convert a node to an object inheriting from NodeRefClass
}
\seealso{
Other Lims Node: \code{\link{LimsRefClass_batchcreate}};
  \code{\link{LimsRefClass_batchretrieve}};
  \code{\link{LimsRefClass_batchupdate}}
}

