% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessRefClass_date_run}
\alias{ProcessRefClass_date_run}
\title{Retrieve the date run as string or POSIXct}
\arguments{
\item{as_POSIX}{logical, by default return a string}
}
\value{
character, POSIXct or "" if not available
}
\description{
Retrieve the date run as string or POSIXct
}
\seealso{
Other Process: \code{\link{ProcessRefClass_get_artifacts}};
  \code{\link{ProcessRefClass_get_parent_process}};
  \code{\link{ProcessRefClass_get_type}};
  \code{\link{ProcessRefClass_technician}};
  \code{\link{create_process_node}}
}

