% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessRefClass_get_artifacts}
\alias{ProcessRefClass_get_artifacts}
\title{Get either the input or the output artifacts (or both!) as NodeRefClass or uri}
\arguments{
\item{what}{either 'input', 'output' or 'both'}

\item{form}{character of either 'Node' or 'uri'}

\item{iom}{an optional data frame of input-output-map data as per 
\code{get_inputoutputmap(form = 'data.frame')}.  If not provided (or NULL)
then this method will fetch it.}
}
\value{
a list of ArtifactRefClass objects or uri unless \code{what} is 
   both in which case a list is returned with 'input' and 'output' elements
}
\description{
Get either the input or the output artifacts (or both!) as NodeRefClass or uri
}
\seealso{
Other Process: \code{\link{ProcessRefClass_date_run}},
  \code{\link{ProcessRefClass_get_file_artifacts}},
  \code{\link{ProcessRefClass_get_parent_process}},
  \code{\link{ProcessRefClass_get_type}},
  \code{\link{ProcessRefClass_instrument}},
  \code{\link{ProcessRefClass_technician}},
  \code{\link{create_process_node}}
}

