% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessRefClass_get_file_artifacts}
\alias{ProcessRefClass_get_file_artifacts}
\title{Retrieve file artifacts (either input, output or both) - very similar to 
get_artifacts but looks for Artifacts types with the keyword 'file' - 'ResultFile',
'SharedResultFile', etc.  Requesting 'output' is fairly quick as the 
input-output-map carries the output-type attribute. Requesting 'input' requires 
more time as each input artifact must be fetched before determining it's type.}
\arguments{
\item{what}{character of 'input', 'output' or 'both'}

\item{form}{character of either 'Node' or 'uri'}

\item{iom}{an optional data frame of input-output-map data as per 
\code{get_inputoutputmap(form = 'data.frame')}.  If not provided (or NULL)
then this method will fetch it.}
}
\value{
a two element list ('input', 'output') of ArtifactRefClass objects or uri 
 depending upon the value of \code{what} one list may be empty
}
\description{
Retrieve file artifacts (either input, output or both) - very similar to 
get_artifacts but looks for Artifacts types with the keyword 'file' - 'ResultFile',
'SharedResultFile', etc.  Requesting 'output' is fairly quick as the 
input-output-map carries the output-type attribute. Requesting 'input' requires 
more time as each input artifact must be fetched before determining it's type.
}
\seealso{
Other Process: \code{\link{ProcessRefClass_date_run}},
  \code{\link{ProcessRefClass_get_artifacts}},
  \code{\link{ProcessRefClass_get_parent_process}},
  \code{\link{ProcessRefClass_get_type}},
  \code{\link{ProcessRefClass_instrument}},
  \code{\link{ProcessRefClass_technician}},
  \code{\link{create_process_node}}
}
