% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessRefClass_get_inputoutputmap}
\alias{ProcessRefClass_get_inputoutputmap}
\title{Get the input-output-maps as InputOutputMapRefClass or a data.frame}
\arguments{
\item{asDataFrame}{logical, if TRUE then return a data.frame}
}
\value{
a list of InputOutputMapRefClass or a data.frame or NULL if not available
}
\description{
Get the input-output-maps as InputOutputMapRefClass or a data.frame
}

