% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessRefClass_get_parent_process}
\alias{ProcessRefClass_get_parent_process}
\title{Get the parent_process assocated with this process}
\arguments{
\item{form}{character flag of type to return "Node" or "uri"}
}
\value{
XML::xmlNode (or NULL) or the uri (or "")
}
\description{
Get the parent_process assocated with this process
}
\seealso{
Other Process: \code{\link{ProcessRefClass_date_run}};
  \code{\link{ProcessRefClass_get_artifacts}};
  \code{\link{ProcessRefClass_get_file_artifacts}};
  \code{\link{ProcessRefClass_get_type}};
  \code{\link{ProcessRefClass_instrument}};
  \code{\link{ProcessRefClass_technician}};
  \code{\link{create_process_node}}
}

