% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessRefClass_get_type}
\alias{ProcessRefClass_get_type}
\title{Retrieve the process type, overrides NodeRefClass_get_type}
\arguments{
\item{form}{'name', or 'uri'}
}
\value{
character or "" if not available
}
\description{
Retrieve the process type, overrides NodeRefClass_get_type
}
\seealso{
Other Process: \code{\link{ProcessRefClass_date_run}};
  \code{\link{ProcessRefClass_get_artifacts}};
  \code{\link{ProcessRefClass_get_file_artifacts}};
  \code{\link{ProcessRefClass_get_parent_process}};
  \code{\link{ProcessRefClass_instrument}};
  \code{\link{ProcessRefClass_technician}};
  \code{\link{create_process_node}}
}

