% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessRefClass_instrument}
\alias{ProcessRefClass_instrument}
\title{Retrieve instrument information (if any)}
\arguments{
\item{form}{'Node', or 'uri'}
}
\value{
InstrumentRefClass, character or NULL if not available
}
\description{
Retrieve instrument information (if any)
}
\seealso{
Other Process: \code{\link{ProcessRefClass_date_run}};
  \code{\link{ProcessRefClass_get_artifacts}};
  \code{\link{ProcessRefClass_get_file_artifacts}};
  \code{\link{ProcessRefClass_get_parent_process}};
  \code{\link{ProcessRefClass_get_type}};
  \code{\link{ProcessRefClass_technician}};
  \code{\link{create_process_node}}
}

