% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessRefClass_technician}
\alias{ProcessRefClass_technician}
\title{Retrieve the technican "First Last" name}
\arguments{
\item{form}{'First Last', "Node" or 'uri'}
}
\value{
ResearcherRefClass, character or NULL if not available
}
\description{
Retrieve the technican "First Last" name
}
\seealso{
Other Process: \code{\link{ProcessRefClass_date_run}},
  \code{\link{ProcessRefClass_get_artifacts}},
  \code{\link{ProcessRefClass_get_file_artifacts}},
  \code{\link{ProcessRefClass_get_parent_process}},
  \code{\link{ProcessRefClass_get_type}},
  \code{\link{ProcessRefClass_instrument}},
  \code{\link{create_process_node}}
}
