% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processtype.R
\name{ProcessTypeRefClass_get_fields}
\alias{ProcessTypeRefClass_get_fields}
\title{Retrieve a named list of fields}
\value{
a named list of FieldRefClass or  NULL
}
\description{
Retrieve a named list of fields
}
\seealso{
Other Processtype: \code{\link{ProcessTypeRefClass_get_inputs}},
  \code{\link{ProcessTypeRefClass_get_outputs}}
}
