% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Processtype.R
\name{ProcessTypeRefClass_get_inputs}
\alias{ProcessTypeRefClass_get_inputs}
\title{Retrieve a named list of process-inputs}
\value{
a named list of inputs NULL
}
\description{
Retrieve a named list of process-inputs
}
\seealso{
Other Processtype: \code{\link{ProcessTypeRefClass_get_fields}};
  \code{\link{ProcessTypeRefClass_get_outputs}}
}

