% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Processtype.R
\name{ProcessTypeRefClass_get_outputs}
\alias{ProcessTypeRefClass_get_outputs}
\title{Retrieve a named list of process-output}
\value{
a named list of outputs NULL
}
\description{
Retrieve a named list of process-output
}
\seealso{
Other Processtype: \code{\link{ProcessTypeRefClass_get_inputs}}
}

