% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{ProjectRefClass_get_attached_files}
\alias{ProjectRefClass_get_attached_files}
\title{Retrieve the attached file associated with a project}
\value{
a list of zero or more FileRefClass or NULL
}
\description{
Retrieve the attached file associated with a project
}
\seealso{
Other Project: \code{\link{ProjectRefClass_get_artifacts}},
  \code{\link{ProjectRefClass_get_containers}},
  \code{\link{ProjectRefClass_get_researcher}},
  \code{\link{ProjectRefClass_get_samples}},
  \code{\link{ProjectRefNode_DELETE}},
  \code{\link{ProjectRefNode_POST}}
}

