% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Project.R
\name{ProjectRefClass_get_containers}
\alias{ProjectRefClass_get_containers}
\title{Retrieve the containers associated with a project}
\arguments{
\item{AA}{optional list of Artifacts in the Project.  If NULL then these
are first retrieved which can be slow.}
}
\value{
a list of zero or more ContainerRefClass or NULL
}
\description{
Retrieve the containers associated with a project
}
\seealso{
Other Project: \code{\link{ProjectRefClass_get_artifacts}};
  \code{\link{ProjectRefClass_get_researcher}};
  \code{\link{ProjectRefClass_get_samples}};
  \code{\link{ProjectRefNode_DELETE}};
  \code{\link{ProjectRefNode_POST}}
}

