% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{ProjectRefClass_get_date}
\alias{ProjectRefClass_get_date}
\title{Retrieve a date "open", "close" or "invoice"}
\arguments{
\item{what}{character either 'open', 'close' or 'invoice'}
}
\value{
a character date %Y-%m-%d or NULL
}
\description{
Retrieve a date "open", "close" or "invoice"
}
