% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{ProjectRefClass_get_researcher}
\alias{ProjectRefClass_get_researcher}
\title{Retrieve the Researcher assigned to the project}
\value{
ResearcherRefClass node or NULL
}
\description{
Retrieve the Researcher assigned to the project
}
\seealso{
Other Project: \code{\link{ProjectRefClass_get_artifacts}},
  \code{\link{ProjectRefClass_get_containers}},
  \code{\link{ProjectRefClass_get_samples}},
  \code{\link{ProjectRefNode_DELETE}},
  \code{\link{ProjectRefNode_POST}}
}

