% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Project.R
\name{ProjectRefClass_get_samples}
\alias{ProjectRefClass_get_samples}
\title{Retrieve the samples associated with a project}
\value{
a list of zero or more SampleRefClass or NULL
}
\description{
This method may take a while depending upon the number of samples in the
system.
}
\seealso{
Other Project: \code{\link{ProjectRefClass_get_artifacts}};
  \code{\link{ProjectRefClass_get_containers}};
  \code{\link{ProjectRefClass_get_researcher}};
  \code{\link{ProjectRefNode_DELETE}};
  \code{\link{ProjectRefNode_POST}}
}

