% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Project.R
\name{ProjectRefNode_DELETE}
\alias{ProjectRefNode_DELETE}
\title{DELETE is disallowed for Project}
\description{
DELETE is disallowed for Project
}
\seealso{
Other Project: \code{\link{ProjectRefClass_get_researcher}};
  \code{\link{ProjectRefNode_POST}}
}

