% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Project.R
\name{ProjectRefNode_POST}
\alias{ProjectRefNode_POST}
\title{POST is disallowed for projects - although see this...
\url{http://genologics.com/files/permanent/API/latest/rest.version.projects.limsid.html}}
\description{
POST is disallowed for projects - although see this...
\url{http://genologics.com/files/permanent/API/latest/rest.version.projects.limsid.html}
}
\seealso{
Other Project: \code{\link{ProjectRefNode_DELETE}}
}

