% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projects.R
\name{ProjectsRefClass_DELETE}
\alias{ProjectsRefClass_DELETE}
\title{DELETE is disabled for ProjectsRefNodes}
\description{
DELETE is disabled for ProjectsRefNodes
}
\seealso{
Other Projects: \code{\link{ProjectsRefClass_PUT}},
  \code{\link{ProjectsRefClass_get_projects_uri}},
  \code{\link{ProjectsRefClass_get_projects}}
}
