% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projects.R
\name{ProjectsRefClass_get_projects}
\alias{ProjectsRefClass_get_projects}
\title{Retrieve a list of projects}
\arguments{
\item{name}{character, one or more names of projects to get - if missing then all}

\item{form}{character, return either 'Node' or 'uri'}
}
\value{
list of ProjectRefClass objects or NULL
}
\description{
Retrieve a list of projects
}
\seealso{
Other Projects: \code{\link{ProjectsRefClass_DELETE}},
  \code{\link{ProjectsRefClass_PUT}},
  \code{\link{ProjectsRefClass_get_projects_uri}}
}

