% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projects.R
\name{ProjectsRefClass_get_projects_uri}
\alias{ProjectsRefClass_get_projects_uri}
\title{Retrieve a list of project uri}
\value{
a named character vector of URIs, possible empty
}
\description{
Retrieve a list of project uri
}
\seealso{
Other Projects: \code{\link{ProjectsRefClass_DELETE}},
  \code{\link{ProjectsRefClass_PUT}},
  \code{\link{ProjectsRefClass_get_projects}}
}
