% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projects.R
\name{ProjectsRefClass_PUT}
\alias{ProjectsRefClass_PUT}
\title{PUT is disabled for ProjectsRefNodes}
\description{
PUT is disabled for ProjectsRefNodes
}
\seealso{
Other Projects: \code{\link{ProjectsRefClass_DELETE}},
  \code{\link{ProjectsRefClass_get_projects_uri}},
  \code{\link{ProjectsRefClass_get_projects}}
}

