% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protocol.R
\docType{class}
\name{ProtocolRefClass-class}
\alias{ProtocolRefClass-class}
\alias{ProtocolRefClass}
\title{A Protocol representation that sublcasses from NodeRefClass}
\description{
A Protocol representation that sublcasses from NodeRefClass
}

