% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_transfer.R
\name{remote_directory_exists}
\alias{remote_directory_exists}
\title{Test if a remote directory exists}
\usage{
remote_directory_exists(uri, username = "foo", password = "bar",
  use = c("cp", "scp", "curl")[2], verbose = FALSE)
}
\arguments{
\item{uri}{character}

\item{username}{the username (required)}

\item{password}{the password}

\item{use}{character one of 'cp', 'scp' or 'curl'}

\item{verbose}{logical, if TRUE then echo the command string}
}
\value{
integer with 0 for success
}
\description{
see https://stackoverflow.com/questions/15927911/how-to-check-if-dir-exists-over-ssh-and-return-results-to-host-machine
}
\seealso{
Other file_transfer: \code{\link{cp_download}},
  \code{\link{duck_download}}, \code{\link{duck_upload}},
  \code{\link{has_duck}}, \code{\link{has_rsync}},
  \code{\link{has_scp}},
  \code{\link{make_remote_directory}},
  \code{\link{rsync_download}}, \code{\link{rsync_upload}},
  \code{\link{scp_download}}, \code{\link{scp_upload}}
}
