% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Researcher.R
\docType{class}
\name{ResearcherRefClass-class}
\alias{ResearcherRefClass}
\alias{ResearcherRefClass-class}
\title{An Researcher representation that sublcasses from NodeRefClass}
\description{
An Researcher representation that sublcasses from NodeRefClass
}
\section{Fields}{

\describe{
\item{\code{character,}}{username the user name}

\item{\code{character,}}{name the user's full name in 'First Last' form}

\item{\code{email}}{character, user's email}

\item{\code{initials}}{character, user's initials}
}}
\seealso{
Other Node: \code{\link{ArtifactGroupRefClass-class}},
  \code{\link{ArtifactRefClass-class}},
  \code{\link{ContainerRefClass-class}},
  \code{\link{ContainerTypeRefClass-class}},
  \code{\link{ExceptionRefClass-class}},
  \code{\link{FieldRefClass-class}},
  \code{\link{FieldRefClass_show}},
  \code{\link{FileRefClass-class}},
  \code{\link{InputOutputMapRefClass-class}},
  \code{\link{NodeRefClass-class}},
  \code{\link{NodeRefClass_BROWSE}},
  \code{\link{NodeRefClass_DELETE}},
  \code{\link{NodeRefClass_GET}},
  \code{\link{NodeRefClass_POST}},
  \code{\link{NodeRefClass_PUT}},
  \code{\link{NodeRefClass_add_namespaces}},
  \code{\link{NodeRefClass_get_childv}},
  \code{\link{NodeRefClass_get_field}},
  \code{\link{NodeRefClass_get_limsid}},
  \code{\link{NodeRefClass_get_type}},
  \code{\link{NodeRefClass_has_child}},
  \code{\link{NodeRefClass_has_lims}},
  \code{\link{NodeRefClass_new}},
  \code{\link{NodeRefClass_required_namespaces}},
  \code{\link{NodeRefClass_set_udfs}},
  \code{\link{NodeRefClass_show}},
  \code{\link{NodeRefClass_toString}},
  \code{\link{NodeRefClass_unames}},
  \code{\link{NodeRefClass_update}},
  \code{\link{ProcessRefClass-class}},
  \code{\link{ProcessTypeRefClass-class}},
  \code{\link{ProjectRefClass-class}},
  \code{\link{ProjectsRefClass-class}},
  \code{\link{SampleRefClass-class}},
  \code{\link{udf_type2R}}
}

