% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Researcher.R
\name{ResearcherRefClass_DELETE}
\alias{ResearcherRefClass_DELETE}
\title{Override the DELETE method to push user to use the GUI}
\value{
NULL invisibly
}
\description{
Override the DELETE method to push user to use the GUI
}
\seealso{
Other Researcher: \code{\link{ResearcherRefClass_POST}};
  \code{\link{ResearcherRefClass_get_credentials}};
  \code{\link{ResearcherRefClass_get_name}};
  \code{\link{ResearcherRefClass_get_username}}
}

