% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Researcher.R
\name{ResearcherRefClass_get_credentials}
\alias{ResearcherRefClass_get_credentials}
\title{Get the credentials as a string.  If the researcher
is uncredentialed then 'none' is returned.}
\value{
character vector
}
\description{
Get the credentials as a string.  If the researcher
is uncredentialed then 'none' is returned.
}
\seealso{
Other Researcher: \code{\link{ResearcherRefClass_DELETE}};
  \code{\link{ResearcherRefClass_POST}};
  \code{\link{ResearcherRefClass_get_name}};
  \code{\link{ResearcherRefClass_get_username}};
  \code{\link{ResearcherRefClass_update}}
}

