% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Researcher.R
\name{ResearcherRefClass_get_lab}
\alias{ResearcherRefClass_get_lab}
\title{Get the lab uri or LabRefClass}
\arguments{
\item{form}{charcater either 'uri' or 'Node'}
}
\value{
charcater uri or LabRefClass
}
\description{
Get the lab uri or LabRefClass
}
\seealso{
Other Researcher: \code{\link{ResearcherRefClass_DELETE}},
  \code{\link{ResearcherRefClass_POST}},
  \code{\link{ResearcherRefClass_get_credentials}},
  \code{\link{ResearcherRefClass_get_name}},
  \code{\link{ResearcherRefClass_get_username}},
  \code{\link{ResearcherRefClass_update}}
}
