% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Researcher.R
\name{ResearcherRefClass_get_username}
\alias{ResearcherRefClass_get_username}
\title{Get the user's username}
\value{
character of username
}
\description{
Get the user's username
}
\seealso{
Other Researcher: \code{\link{ResearcherRefClass_DELETE}},
  \code{\link{ResearcherRefClass_POST}},
  \code{\link{ResearcherRefClass_get_credentials}},
  \code{\link{ResearcherRefClass_get_name}},
  \code{\link{ResearcherRefClass_update}}
}

