% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Researcher.R
\name{ResearcherRefClass_POST}
\alias{ResearcherRefClass_POST}
\title{Override the POST method to push user to use the GUI}
\value{
NULL invisibly
}
\description{
Override the POST method to push user to use the GUI
}
\seealso{
Other Researcher: \code{\link{ResearcherRefClass_DELETE}};
  \code{\link{ResearcherRefClass_get_credentials}};
  \code{\link{ResearcherRefClass_get_name}};
  \code{\link{ResearcherRefClass_get_username}};
  \code{\link{ResearcherRefClass_update}}
}

