% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Researcher.R
\name{ResearcherRefClass_update}
\alias{ResearcherRefClass_update}
\title{Update the object fields}
\arguments{
\item{...}{further arguments for \code{NodeRefClass$update()}}
}
\description{
Update the object fields
}
\seealso{
Other Researcher: \code{\link{ResearcherRefClass_DELETE}},
  \code{\link{ResearcherRefClass_POST}},
  \code{\link{ResearcherRefClass_get_credentials}},
  \code{\link{ResearcherRefClass_get_lab}},
  \code{\link{ResearcherRefClass_get_name}},
  \code{\link{ResearcherRefClass_get_username}}
}
