% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file_transfer.R
\name{rsync_download}
\alias{rsync_download}
\title{Download a file using rsync - rsync and then rename}
\usage{
rsync_download(url, dest, username = "foo", password = "bar",
  verbose = FALSE, extra = "-q")
}
\arguments{
\item{url}{character}

\item{dest}{character destination filename, by default the basename of the URL}

\item{username}{the username (required)}

\item{password}{the password (required)}

\item{verbose}{logical, if TRUE then echo the command string}

\item{extra}{character extra params for scp, currently "-q"}
}
\value{
integer with 0 for success
}
\description{
This assumes that SSH public-key authentication is set up.
}
\seealso{
Other file_transfer: \code{\link{cp_download}};
  \code{\link{duck_download}}; \code{\link{duck_upload}};
  \code{\link{has_duck}}; \code{\link{has_rsync}};
  \code{\link{has_scp}}; \code{\link{rsync_upload}};
  \code{\link{scp_download}}; \code{\link{scp_upload}}
}

