% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Sample.R
\name{SampleRefClass_get_all_artifacts}
\alias{SampleRefClass_get_all_artifacts}
\title{Get ALL associated artifacts.  This can be painfully slow and probably is
not worth doing.}
\value{
list of ArtifactRefClass or NULL
}
\description{
Get ALL associated artifacts.  This can be painfully slow and probably is
not worth doing.
}
\seealso{
Other Sample: \code{\link{SampleRefClass_get_artifact}};
  \code{\link{SampleRefClass_get_biosource}};
  \code{\link{SampleRefClass_get_container}};
  \code{\link{SampleRefClass_get_project}};
  \code{\link{SampleRefNode_DELETE}};
  \code{\link{SampleRefNode_POST}}
}

