% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Sample.R
\name{SampleRefClass_get_biosource}
\alias{SampleRefClass_get_biosource}
\title{Get Biosource description}
\value{
character, possible empty
}
\description{
Get Biosource description
}
\seealso{
Other Sample: \code{\link{SampleRefClass_get_artifact}};
  \code{\link{SampleRefClass_get_container}};
  \code{\link{SampleRefClass_get_project}};
  \code{\link{SampleRefNode_DELETE}};
  \code{\link{SampleRefNode_POST}}
}

