% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Sample.R
\name{SampleRefClass_get_container}
\alias{SampleRefClass_get_container}
\title{Get the sample container as uri or NodeRefClass}
\arguments{
\item{form}{character either "uri" or "Node"}
}
\value{
ContainerRefClass or uri (or NULL or "")
}
\description{
Get the sample container as uri or NodeRefClass
}
\seealso{
Other Sample: \code{\link{SampleRefClass_get_artifact}};
  \code{\link{SampleRefClass_get_biosource}};
  \code{\link{SampleRefClass_get_project}};
  \code{\link{SampleRefNode_DELETE}};
  \code{\link{SampleRefNode_POST}}
}

