% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.R
\name{SampleRefClass_get_location}
\alias{SampleRefClass_get_location}
\title{Get location as 'containerName_placement'}
\arguments{
\item{...}{further arguments for ArtifactRefClass_get_location}

\item{character}{string, possibly empty}
}
\description{
Get location as 'containerName_placement'
}
\seealso{
Other Sample: \code{\link{SampleRefClass_get_all_artifacts}},
  \code{\link{SampleRefClass_get_artifact}},
  \code{\link{SampleRefClass_get_biosource}},
  \code{\link{SampleRefClass_get_container}},
  \code{\link{SampleRefClass_get_project}},
  \code{\link{SampleRefNode_DELETE}},
  \code{\link{SampleRefNode_POST}}
}
