% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Sample.R
\name{SampleRefClass_get_project}
\alias{SampleRefClass_get_project}
\title{Get project as uri or Node}
\arguments{
\item{form}{character either "Node" or "uri"}
}
\value{
character of NodeRefClass, possibly "" or NULL
}
\description{
Get project as uri or Node
}
\seealso{
Other Sample: \code{\link{SampleRefClass_get_all_artifacts}};
  \code{\link{SampleRefClass_get_artifact}};
  \code{\link{SampleRefClass_get_biosource}};
  \code{\link{SampleRefClass_get_container}};
  \code{\link{SampleRefNode_DELETE}};
  \code{\link{SampleRefNode_POST}}
}

