% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Sample.R
\name{SampleRefNode_POST}
\alias{SampleRefNode_POST}
\title{POST is disallowed for samples}
\description{
POST is disallowed for samples
}
\seealso{
Other Sample: \code{\link{SampleRefClass_get_artifact}};
  \code{\link{SampleRefClass_get_biosource}};
  \code{\link{SampleRefClass_get_project}};
  \code{\link{SampleRefNode_DELETE}}
}

