% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file_transfer.R
\name{scp_upload}
\alias{scp_upload}
\title{Upload a file using scp.}
\usage{
scp_upload(filename, url, username = "foo", password = "bar")
}
\arguments{
\item{filename}{the fully qualified filename of file to upload}

\item{url}{the destination url}

\item{username}{the username (required)}

\item{password}{the password (required)}
}
\value{
integer 0 for success
}
\description{
This assumes that SSH public-key authentication is set up.
}
\seealso{
Other file_transfer: \code{\link{cp_download}};
  \code{\link{duck_download}}; \code{\link{duck_upload}};
  \code{\link{has_duck}}; \code{\link{has_rsync}};
  \code{\link{has_scp}}; \code{\link{rsync_download}};
  \code{\link{rsync_upload}}; \code{\link{scp_download}}
}

