% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Udf.R
\name{set_udfs}
\alias{set_udfs}
\title{Set one or more UDF fields in an xmlNode}
\usage{
set_udfs(x, v)
}
\arguments{
\item{x}{XML::xmlNode}

\item{v}{a list of one or more udf vectors
each udf vector must have \code{name}, \code{type} and \code{value}}
}
\value{
the updated xmlNode
}
\description{
Set one or more UDF fields in an xmlNode
}
\seealso{
Other Node Udf: \code{\link{NodeRefClass_get_udfs}};
  \code{\link{extract_udfs}}; \code{\link{get_udfs}}
}

