% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{singular}
\alias{singular}
\title{Convert to singular from plural}
\usage{
singular(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
Convert to singular from plural
}

