% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{split_vector}
\alias{split_vector}
\title{Split a vector into groups of MAX (or possibly fewer)}
\usage{
split_vector(v, MAX = 200)
}
\arguments{
\item{v}{vector or list to split}

\item{MAX}{numeric the maximum size per group}
}
\value{
a list of the vector split into groups
}
\description{
Split a vector into groups of MAX (or possibly fewer)
}

