% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage.R
\docType{class}
\name{StageRefClass-class}
\alias{StageRefClass-class}
\alias{StageRefClass}
\title{A Stage representation that sublcasses from NodeRefClass}
\description{
A Stage representation that sublcasses from NodeRefClass
}

