% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step.R
\docType{class}
\name{StepRefClass-class}
\alias{StepRefClass-class}
\alias{StepRefClass}
\title{A Step representation that sublcasses from NodeRefClass}
\description{
A Step representation that sublcasses from NodeRefClass
}

