% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{trimuri}
\alias{trimuri}
\title{Split a uri around the "?" or other character.  For example, 
\code{trimuri(http://www.uri.org?where=foo)} returns \code{http://www.uri.org}.}
\usage{
trimuri(uri, around = "?", fixed = TRUE, index = 1)
}
\arguments{
\item{uri}{character, the uri to split}

\item{around}{character,the character to split around}

\item{fixed}{see \code{\link{strsplit}}}

\item{index}{numeric, by default only the leading portion of the input is
returned}
}
\value{
character of the input preceding the 'around' character
}
\description{
Split a uri around the "?" or other character.  For example, 
\code{trimuri(http://www.uri.org?where=foo)} returns \code{http://www.uri.org}.
}

