% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{try_GET}
\alias{try_GET}
\title{Get a uri with option to retry up to \code{tries} times.  Useful when doing
vulnerable batch operations but could be used anytime.}
\usage{
try_GET(uri, lims, ..., tries = 3)
}
\arguments{
\item{uri}{character, the uri to retrieve}

\item{lims}{a LimsRefClass node}

\item{...}{further arguments for httr::GET()}

\item{tries}{numeric, allow up to this number of tries before failing}
}
\value{
result of httr::GET
}
\description{
Get a uri with option to retry up to \code{tries} times.  Useful when doing
vulnerable batch operations but could be used anytime.
}

