% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{unames}
\alias{unames}
\title{Return a vector of unique names from a objeect.  The object must have the
methods for \code{names}}
\usage{
unames(x)
}
\arguments{
\item{x}{the object}
}
\value{
a vector of unqiue names or NULL
}
\description{
Return a vector of unique names from a objeect.  The object must have the
methods for \code{names}
}

