% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\docType{class}
\name{WorkflowRefClass-class}
\alias{WorkflowRefClass-class}
\alias{WorkflowRefClass}
\title{A Workflow representation that sublcasses from NodeRefClass}
\description{
A Workflow representation that sublcasses from NodeRefClass
}

