% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\name{WorkflowRefClass_get_stages}
\alias{WorkflowRefClass_get_stages}
\title{Get one or more stages}
\arguments{
\item{name}{NULL or charater vector of stage names to retrieve}

\item{form}{character either 'uri' or 'Node'}
}
\value{
vector of URI, list of Nodes or empty list/vector
}
\description{
Get one or more stages
}
