% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{xml_atts}
\alias{xml_atts}
\title{Extract the attributes from a simple XML::xmlNode object}
\usage{
xml_atts(x, ...)
}
\arguments{
\item{x}{XML::xmlNode with attributes}

\item{...}{further arguments for \code{XML::xmlAttrs()}}

\item{character}{vector of the attributes}
}
\description{
Extract the attributes from a simple XML::xmlNode object
}

