% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{xml_name}
\alias{xml_name}
\title{Extract the name from a simple XML::xmlNode object}
\usage{
xml_name(x, ...)
}
\arguments{
\item{x}{XML::xmlNode with a value}

\item{...}{further arguments for \code{XML::xmlName()}}
}
\value{
the name of the node
}
\description{
Extract the name from a simple XML::xmlNode object
}

