% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{xml_string}
\alias{xmlString}
\alias{xml_string}
\title{Convert XML::xmlNode or NodeRefClass to character}
\usage{
xml_string(x)
}
\arguments{
\item{x}{xmlNode or NodeRefClass}
}
\value{
character
}
\description{
Convert XML::xmlNode or NodeRefClass to character
}

