% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{xmlString}
\alias{xmlString}
\title{Convert XML::xmlNode or NodeRefClass to character}
\usage{
xmlString(x, replace_newline = TRUE)
}
\arguments{
\item{x}{xmlNode or NodeRefClass}

\item{replace_newline}{logical, if TRUE replace \code{\n} with ""}
}
\value{
character
}
\description{
Convert XML::xmlNode or NodeRefClass to character
}
