% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Artifact.R
\name{ArtifactRefClass_working_flag}
\alias{ArtifactRefClass_working_flag}
\title{Get the working flag}
\value{
charcater of containerLimsid_location
}
\description{
Get the working flag
}
\seealso{
Other Artifact: \code{\link{ArtifactRefClass_get_container}};
  \code{\link{ArtifactRefClass_get_file_artifact}};
  \code{\link{ArtifactRefClass_get_parent_process}};
  \code{\link{ArtifactRefClass_get_sample}};
  \code{\link{ArtifactRefClass_get_wellname}};
  \code{\link{ArtifactRefClass_location}};
  \code{\link{ArtifactRefClass_qc_flag}};
  \code{\link{ArtifactRefNode_DELETE}};
  \code{\link{ArtifactRefNode_POST}}
}

