% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{batch_retrieve}
\alias{batch_retrieve}
\title{Retrieve a batch resource}
\usage{
batch_retrieve(uri, lims, rel = c("artifacts", "containers", "files",
  "samples")[1], resource = file.path(rel, "batch", "retrieve"),
  asList = TRUE, rm_dups = TRUE, ...)
}
\arguments{
\item{uri}{character vector of one or more uri}

\item{lims}{LimsRefClass object}

\item{resource}{character}

\item{asList}{logical, if TRUE return a named list of NodeRefClass objects}

\item{...}{further arguments}

\item{relative}{resource name}

\item{all}{logical, by default we remove duplicates set this to TRUE to 
retrieve all, ignored if \code{asList = FALSE}}
}
\value{
a list of NodeRefClass
}
\description{
Retrieve a batch resource
}

