% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefClass_get_artifacts}
\alias{ContainerRefClass_get_artifacts}
\title{Retrieve a named list ArtifactRefClass objects}
\value{
a named list of ArtifactRefClass objects
}
\description{
Retrieve a named list ArtifactRefClass objects
}
\seealso{
Other Container: \code{\link{ContainerRefClass_POST}};
  \code{\link{ContainerRefClass_get_placements}};
  \code{\link{ContainerRefClass_get_samples}};
  \code{\link{ContainerRefClass_n_empty}};
  \code{\link{ContainerRefClass_n_occupied}};
  \code{\link{ContainerRefClass_n_wells}}
}

