% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefClass_get_samples}
\alias{ContainerRefClass_get_samples}
\title{Retrieve a named list of artifact XML:xmlNode or SampleRefClass}
\arguments{
\item{placement}{well name such as 'A:1' or NULL to retrieve all}

\item{asNode}{logical if TRUE cast the result to SampleRefClass}
}
\value{
a named vector of sample XML:xmlNode or SampleRefClass objects
}
\description{
Retrieve a named list of artifact XML:xmlNode or SampleRefClass
}
\seealso{
Other Container: \code{\link{ContainerRefClass_POST}};
  \code{\link{ContainerRefClass_get_artifacts}};
  \code{\link{ContainerRefClass_get_placements}};
  \code{\link{ContainerRefClass_n_empty}};
  \code{\link{ContainerRefClass_n_occupied}};
  \code{\link{ContainerRefClass_n_wells}}
}

