% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container.R
\name{ContainerRefClass_n_wells}
\alias{ContainerRefClass_n_wells}
\title{Retrieve the well count (occupied + empty)}
\value{
numeric
}
\description{
Retrieve the well count (occupied + empty)
}
\seealso{
Other Container: \code{\link{ContainerRefClass_POST}},
  \code{\link{ContainerRefClass_get_artifacts}},
  \code{\link{ContainerRefClass_get_placements}},
  \code{\link{ContainerRefClass_get_samples}},
  \code{\link{ContainerRefClass_get_type}},
  \code{\link{ContainerRefClass_n_empty}},
  \code{\link{ContainerRefClass_n_occupied}},
  \code{\link{ContainerRefClass_required_namespaces}}
}

