% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_transfer.R
\name{duck_download}
\alias{duck_download}
\title{Download a file using duck}
\usage{
duck_download(url, dest, username = "foo", password = "bar",
  verbose = FALSE, extra = "--quiet --existing overwrite")
}
\arguments{
\item{url}{the url of the file}

\item{dest}{character destination filename, by default the basename of the URL}

\item{username}{the username (required)}

\item{password}{the password (required)}

\item{verbose}{logical, if TRUE then echo the command string}

\item{extra}{character extra params for duck, currently "--existing overwrite"}
}
\value{
integer with 0 for success
}
\description{
Download a file using duck
}
\seealso{
Other file_transfer: \code{\link{cp_download}},
  \code{\link{duck_upload}}, \code{\link{has_duck}},
  \code{\link{has_rsync}}, \code{\link{has_scp}},
  \code{\link{rsync_download}}, \code{\link{rsync_upload}},
  \code{\link{scp_download}}, \code{\link{scp_upload}}
}
