% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/File.R
\name{FileRefClass_download}
\alias{FileRefClass_download}
\title{Download the file identified in 'content-location'}
\arguments{
\item{dest}{filename for destination, by default the basename of the 'content_location'}

\item{use}{character indicating download tool to use.  Currently just 'duck'}

\item{up}{character vector of two elements [username, password].  If not
provided the then the credentials are derived from the File nodes' Lims.}

\item{...}{further arguments for the download tool}
}
\value{
numeric code where 0 means success
}
\description{
Download tools include
\itemize{
   \item{duck}{From Cyberduck \url{https://duck.sh/}}
   \item{scp}{not implemented yet}
   \item{curl}{not implemented yet}
}
}

