% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genologicsr-package.R
\docType{package}
\name{genologicsr}
\alias{genologicsr}
\alias{genologicsr-package}
\title{genologicsr-package.R
A simplified interface between XML nodes and the Genologics RESTful API.}
\description{
genologicsr-package.R
A simplified interface between XML nodes and the Genologics RESTful API.
}

