% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genologicsr_package.R
\docType{package}
\name{genologicsr}
\alias{genologicsr}
\alias{genologicsr-package}
\title{genologicsr: R wrapper around the Genologics API}
\description{
genologics provides a simplified interface between XML nodes and the 
Genologics RESTful API.
}
