% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exception.R
\name{is_exception}
\alias{is_exception}
\title{Test XML::xmlNode or NodeRefClass is an exception}
\usage{
is_exception(x, space = "exc", include_null = TRUE)
}
\arguments{
\item{x}{object to test}

\item{space}{the namespace to test}

\item{include_null}{logical, if TRUE assume that NULL is an exception}
}
\value{
logical
}
\description{
Test XML::xmlNode or NodeRefClass is an exception
}
