% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{Lims}
\alias{Lims}
\title{Instantiate a LimsRefClass object}
\usage{
Lims(configfile = build_config_path())
}
\arguments{
\item{configfile}{character, the fully qualified path to the config file}
}
\value{
a LimsRefClass instance or NULL
}
\description{
Instantiate a LimsRefClass object
}
