% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_check}
\alias{LimsRefClass_check}
\title{Verify a response and return xmlNode, possible an exception node}
\arguments{
\item{rsp}{httr::response object}

\item{msg}{if an exception is encountered, attach this message (unless NULL)}
}
\value{
XML::xmlNode
}
\description{
Verify a response and return xmlNode, possible an exception node
}
\seealso{
Other Lims: \code{\link{LimsRefClass-class}},
  \code{\link{LimsRefClass_BROWSE}},
  \code{\link{LimsRefClass_DELETE}},
  \code{\link{LimsRefClass_POST}},
  \code{\link{LimsRefClass_PUSH}},
  \code{\link{LimsRefClass_PUT}},
  \code{\link{LimsRefClass_get_byLimsid}},
  \code{\link{LimsRefClass_uri}},
  \code{\link{LimsRefClass_userpwd}},
  \code{\link{LimsRefClass_validate_session}},
  \code{\link{get_NSMAP}}
}

