% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_create_exception}
\alias{LimsRefClass_create_exception}
\title{Create an exception node}
\arguments{
\item{message}{chracter, some error message}

\item{asNode}{logical if TRUE return ExceptionRefClass}
}
\value{
XML::xmlNode of ExceptionRefClass
}
\description{
Create an exception node
}

