% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lims.R
\name{LimsRefClass_DELETE}
\alias{LimsRefClass_DELETE}
\title{DELETE a resource}
\arguments{
\item{x}{XML::xmlNode to DELETE, generally a file node}

\item{...}{further arguments for httr::DELETE}
}
\value{
logical
}
\description{
Typically this is a file (resource = 'files')
}
\seealso{
Other Lims: \code{\link{LimsRefClass_POST}};
  \code{\link{LimsRefClass_PUSH}};
  \code{\link{LimsRefClass_PUT}};
  \code{\link{LimsRefClass_get_byLimsid}};
  \code{\link{LimsRefClass_uri}};
  \code{\link{LimsRefClass_validate_session}};
  \code{\link{LimsRefClass}},
  \code{\link{LimsRefClass-class}}; \code{\link{get_NSMAP}}
}

