% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{read_config}
\alias{read_config}
\title{Read a configuration file}
\usage{
read_config(filename)
}
\arguments{
\item{filename}{the name of the file}
}
\value{
a name list with one element per section
 each section, in turn is a named list
}
\description{
Read a configuration file
}

