% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Researcher.R
\docType{class}
\name{ResearcherRefClass-class}
\alias{ResearcherRefClass}
\alias{ResearcherRefClass-class}
\title{An Researcher representation that sublcasses from NodeRefClass}
\description{
An Researcher representation that sublcasses from NodeRefClass
}
\section{Fields}{

\describe{
\item{\code{character,}}{username the user name}

\item{\code{character,}}{name the user's full name in 'First Last' form}

\item{\code{email}}{character, user's email}

\item{\code{initials}}{character, user's initials}
}}
\seealso{
Other Node: \code{\link{ArtifactRefClass}},
  \code{\link{ArtifactRefClass-class}};
  \code{\link{ContainerRefClass}},
  \code{\link{ContainerRefClass-class}};
  \code{\link{FileRefClass}},
  \code{\link{FileRefClass-class}};
  \code{\link{InputOutputMapRefClass}},
  \code{\link{InputOutputMapRefClass-class}};
  \code{\link{NodeRefClass_DELETE}};
  \code{\link{NodeRefClass_GET}};
  \code{\link{NodeRefClass_POST}};
  \code{\link{NodeRefClass_PUT}};
  \code{\link{NodeRefClass_get_field}};
  \code{\link{NodeRefClass_get_type}},
  \code{\link{NodeRefClass_get_type}};
  \code{\link{NodeRefClass_has_child}};
  \code{\link{NodeRefClass_has_lims}};
  \code{\link{NodeRefClass_new}};
  \code{\link{NodeRefClass_show}};
  \code{\link{NodeRefClass_toString}};
  \code{\link{NodeRefClass_unames get the unique child
  names}}; \code{\link{NodeRefClass_update}};
  \code{\link{NodeRefClass}},
  \code{\link{NodeRefClass-class}};
  \code{\link{ProcessRefClass}},
  \code{\link{ProcessRefClass-class}};
  \code{\link{SampleRefClass}},
  \code{\link{SampleRefClass-class}};
  \code{\link{udf_type2R}}
}

