% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Researcher.R
\name{ResearcherRefClass_get_name}
\alias{ResearcherRefClass_get_name}
\title{Get the user's First Last name (overrides NodeRefClass_get_name)}
\value{
character of First Last
}
\description{
Get the user's First Last name (overrides NodeRefClass_get_name)
}
\seealso{
Other Researcher: \code{\link{ResearcherRefClass_DELETE}};
  \code{\link{ResearcherRefClass_POST}};
  \code{\link{ResearcherRefClass_get_credentials}};
  \code{\link{ResearcherRefClass_get_username}};
  \code{\link{ResearcherRefClass_update}}
}

