% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.R
\name{SampleRefNode_DELETE}
\alias{SampleRefNode_DELETE}
\title{DELETE is disallowed for samples}
\description{
DELETE is disallowed for samples
}
\seealso{
Other Sample: \code{\link{SampleRefClass_get_all_artifacts}},
  \code{\link{SampleRefClass_get_artifact}},
  \code{\link{SampleRefClass_get_biosource}},
  \code{\link{SampleRefClass_get_container}},
  \code{\link{SampleRefClass_get_location}},
  \code{\link{SampleRefClass_get_project}},
  \code{\link{SampleRefNode_POST}}
}
