% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Udf.R
\name{set_udfs}
\alias{set_udfs}
\title{Set one or more UDF fields in an xmlNode.}
\usage{
set_udfs(x, v)
}
\arguments{
\item{x}{XML::xmlNode}

\item{v}{a list of one or more udf vectors
each udf vector must have \code{name}, \code{type} and \code{value}}
}
\value{
the updated xmlNode
}
\description{
Non-Ascii non-UTF8 characters
 are scrubbed from the input values.
}

